Calculate Coefficient of Conservatism metrics

This function calculates C of C metrics, including variations based on native status, if NWCA_NATSTAT is present in the input data frame.

calcCC(vascIn, sampID = "UID")

Arguments

vascIn

Data frame containing cover data summarized by UID and TAXON, with the following fields:

sampID: Variable(s) identified in sampID argument

TAXON: Taxon name

XABCOV: Mean percent cover of taxon across plots

FREQ: Number plots in which taxon occurs

NWCA_CC: Coefficient of conservatism values by taxon from NWCA

NWCA_NATSTAT (optional): Native status variable with categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'. UND taxa are ignored.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing sampID variables, PARAMETER, RESULT, where values of PARAMETER are:

XC: Mean coefficient of conservatism (unweighted)

FQAI: Floral Quality Assessment Index (unweighted)

XC_FREQ: Mean coefficient of conservatism weighted by relative frequency

FQAI_FREQ: Floral Quality Assessment Index weighted by relative frequency

XC_COV: Mean coefficient of conservatism weighted by relative absolute cover

FQAI_COV: Floral Quality Assessment Index weighted by relative absolute cover

For metrics based native species, the metric name has a suffix of NAT.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom Blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

ccEx <- calcCC(exPlant$byUIDspp)

head(ccEx)
#>           UID PARAMETER RESULT
#> 9.XC_ALL    9    XC_ALL   2.23
#> 2.XC_ALL    2    XC_ALL   4.49
#> 3.XC_ALL    3    XC_ALL   4.67
#> 6.XC_ALL    6    XC_ALL   4.77
#> 1.XC_ALL    1    XC_ALL   4.78
#> 10.XC_ALL  10    XC_ALL   4.90
unique(ccEx$PARAMETER)
#>  [1] "XC_ALL"        "XC_FREQ_ALL"   "XC_COV_ALL"    "FQAI_ALL"     
#>  [5] "FQAI_FREQ_ALL" "FQAI_COV_ALL"  "N_SEN"         "XRCOV_SEN"    
#>  [9] "XABCOV_SEN"    "PCTN_SEN"      "N_TOL"         "XRCOV_TOL"    
#> [13] "XABCOV_TOL"    "PCTN_TOL"      "N_ISEN"        "XRCOV_ISEN"   
#> [17] "XABCOV_ISEN"   "PCTN_ISEN"     "N_HTOL"        "XRCOV_HTOL"   
#> [21] "XABCOV_HTOL"   "PCTN_HTOL"     "N_HSEN"        "XRCOV_HSEN"   
#> [25] "XABCOV_HSEN"   "PCTN_HSEN"     "XC_NAT"        "XC_FREQ_NAT"  
#> [29] "XC_COV_NAT"    "FQAI_NAT"      "FQAI_FREQ_NAT" "FQAI_COV_NAT"