Calculate plant category metrics

This function calculates all category-based metrics, including versions with only native species if the variable NWCA_NATSTAT is included in the input data frame.

calcCategory(vascIn, sampID = "UID")

Arguments

vascIn

Data frame containing cover data summarized by UID and TAXON, with the following fields:

sampID: Variable(s) identified in sampID argument

TAXON: Taxon name

CATEGORY: USDA PLANTS category variable, with valid values of DICOT, FERN, GYMNOSPERM, HORSETAIL, LIVERWORT, LYCOPOD, MONOCOT, QUILLWORT, or blank

XABCOV: Mean percent cover of taxon across plots

TOTN: Number of taxa in sample

sXRCOV: proportion of summed cover across all taxa (XTOTABCOV) represented by taxon in sample

NWCA_NATSTAT (optional): Native status variable with categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait value (represented as TRAITNM below):

N_TRAITNM: Number of taxa with trait

PCTN_TRAITNM: Number of taxa with trait as percentage of TOTN

XABCOV_TRAITNM: Sum of XABCOV values across taxa with trait

XRCOV_TRAITNM: Sum of sXRCOV values across taxa with trait

For metrics using native status, the metric name has a suffix of NAT, ALIEN, AC, INTR, or CRYP.

Details

Both CATEGORY and NWCA_NATSTAT variables are recoded to fewer categories. Taxa with 'UND' as native status are excluded.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom Blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

catEx <- calcCategory(exPlant$byUIDspp)

head(catEx)
#>            UID PARAMETER RESULT
#> 8.N_DICOT    8   N_DICOT      0
#> 5.N_DICOT    5   N_DICOT      2
#> 7.N_DICOT    7   N_DICOT      3
#> 10.N_DICOT  10   N_DICOT      5
#> 3.N_DICOT    3   N_DICOT      8
#> 9.N_DICOT    9   N_DICOT     11
unique(catEx$PARAMETER)
#>  [1] "N_DICOT"               "N_FERN"                "N_MONOCOT"            
#>  [4] "N_GYMNOSPERM"          "N_HORSETAIL"           "PCTN_DICOT"           
#>  [7] "PCTN_FERN"             "PCTN_MONOCOT"          "PCTN_GYMNOSPERM"      
#> [10] "PCTN_HORSETAIL"        "XABCOV_DICOT"          "XABCOV_FERN"          
#> [13] "XABCOV_MONOCOT"        "XABCOV_GYMNOSPERM"     "XABCOV_HORSETAIL"     
#> [16] "XRCOV_DICOT"           "XRCOV_FERN"            "XRCOV_MONOCOT"        
#> [19] "XRCOV_GYMNOSPERM"      "XRCOV_HORSETAIL"       "N_DICOTS_NAT"         
#> [22] "XRCOV_DICOTS_NAT"      "XABCOV_DICOTS_NAT"     "PCTN_DICOTS_NAT"      
#> [25] "N_DICOTS_ALIEN"        "PCTN_DICOTS_ALIEN"     "XABCOV_DICOTS_ALIEN"  
#> [28] "XRCOV_DICOTS_ALIEN"    "N_DICOTS_CRYP"         "PCTN_DICOTS_CRYP"     
#> [31] "XABCOV_DICOTS_CRYP"    "XRCOV_DICOTS_CRYP"     "N_DICOTS_AC"          
#> [34] "PCTN_DICOTS_AC"        "XABCOV_DICOTS_AC"      "XRCOV_DICOTS_AC"      
#> [37] "N_FERNS_NAT"           "XRCOV_FERNS_NAT"       "XABCOV_FERNS_NAT"     
#> [40] "PCTN_FERNS_NAT"        "N_FERNS_INTR"          "PCTN_FERNS_INTR"      
#> [43] "XABCOV_FERNS_INTR"     "XRCOV_FERNS_INTR"      "N_MONOCOTS_NAT"       
#> [46] "XRCOV_MONOCOTS_NAT"    "XABCOV_MONOCOTS_NAT"   "PCTN_MONOCOTS_NAT"    
#> [49] "N_MONOCOTS_ALIEN"      "PCTN_MONOCOTS_ALIEN"   "XABCOV_MONOCOTS_ALIEN"
#> [52] "XRCOV_MONOCOTS_ALIEN"  "N_MONOCOTS_CRYP"       "PCTN_MONOCOTS_CRYP"   
#> [55] "XABCOV_MONOCOTS_CRYP"  "XRCOV_MONOCOTS_CRYP"   "N_MONOCOTS_AC"        
#> [58] "PCTN_MONOCOTS_AC"      "XABCOV_MONOCOTS_AC"    "XRCOV_MONOCOTS_AC"    
#> [61] "N_LYCOPOD"             "N_QUILLWORT"           "N_LIVERWORT"          
#> [64] "PCTN_LYCOPOD"          "PCTN_QUILLWORT"        "PCTN_LIVERWORT"       
#> [67] "XABCOV_LYCOPOD"        "XABCOV_QUILLWORT"      "XABCOV_LIVERWORT"     
#> [70] "XRCOV_LYCOPOD"         "XRCOV_QUILLWORT"       "XRCOV_LIVERWORT"