Calculate diversity indices

Calculate Simpson Diversity, Shannon-Wiener Diversity, and Pielou's Evenness indices, with variations based on native status if NWCA_NATSTAT is included in the input data frame.

calcDiversity(vascIn, sampID = "UID")

Arguments

vascIn

Data frame containing cover data summarized by UID and TAXON, with the following fields:

sampID: Variable(s) identified in sampID argument

TAXON: Taxon name

CATEGORY: USDA PLANTS category variable

XABCOV: Mean percent cover of taxon across plots

Optional: NWCA_NATSTAT: Native status variable with categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing sampID variables, PARAMETER, RESULT, where values of PARAMETER are:

D_ALL: Simpson diversity index based on all taxa

H_ALL: Shannon-Wiener diversity index based on all taxa

J_ALL: Pielou's evenness index based on all taxa

If NWCA_NATSTAT is included in input data frame, the following variations based on subsets of taxa are calculated:

D_NAT: Simpson diversity based only on native species

D_ALIEN: Simpson diversity based only on alien species

D_AC: Simpson diversity based only on alien and cryptogenic species

H_NAT: Shannon-Wiener diversity index based on native species

H_ALIEN: Shannon-Wiener diversity index based on alien species

H_AC: Shannon-Wiener diversity index based on alien and cryptogenic species

J_NAT: Pielou's evenness index based on native taxa

J_ALIEN: Pielou's evenness index based on alien taxa

J_AC: Pielou's evenness index based on alien and cryptogenic taxa

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

divEx <- calcDiversity(exPlant$byUIDspp)

head(divEx)
#>         UID PARAMETER RESULT
#> 1.H_ALL   1     H_ALL 1.9476
#> 2.H_ALL   2     H_ALL 2.0566
#> 3.H_ALL   3     H_ALL 1.6998
#> 4.H_ALL   4     H_ALL 3.1242
#> 5.H_ALL   5     H_ALL 0.6146
#> 6.H_ALL   6     H_ALL 2.6861
unique(divEx$PARAMETER)
#>  [1] "H_ALL"   "J_ALL"   "D_ALL"   "H_NAT"   "J_NAT"   "D_NAT"   "H_ALIEN"
#>  [8] "J_ALIEN" "D_ALIEN" "H_AC"    "J_AC"    "D_AC"