Calculate vascular plant duration metrics

This function calculates all duration metrics, with additional metrics if NWCA_NATSTAT variable is included in input data frame.

calcDuration(vascIn, sampID = "UID")

Arguments

vascIn

Data frame containing cover data summarized by sampID variables and TAXON, with the following fields:

sampID: Variable(s) in the argument sampID

TAXON: Taxon name

DURATION: USDA PLANTS DURATION variable, with valid values "ANNUAL", "BIENNIAL", "PERENNIAL", combinations of these values, or blank

DUR_ALT (optional): Combinations of DURATION into ANNUAL, PERENNIAL, ANN_BIEN, and ANN_PEREN, as used in NWCA. Created in code if not supplied by user.

XABCOV: Mean percent cover of taxon across plots

TOTN: Number of taxa in sample

sXRCOV: proportion of summed cover across all taxa (XTOTABCOV) represented by taxon in sample

NWCA_NATSTAT (Optional): Native status variable with categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing the sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait value (represented as TRAITNM below):

N_TRAITNM: Number of taxa with trait

PCTN_TRAITNM: Number of taxa with trait as percentage of TOTN

XABCOV_TRAITNM: Sum of XABCOV values across taxa with trait

XRCOV_TRAITNM: Sum of sXRCOV values across taxa with trait

If NWCA_NATSTAT is in the input data frame, the same set except 'N_' metrics is calculated for native species and alien + cryptogenic species, with metric names suffixes of _NAT and _AC, respectively.

Details

Both DURATION and NWCA_NATSTAT variables are recoded to fewer categories. Taxa with 'UND' as native status are excluded.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom Blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

durEx <- calcDuration(exPlant$byUIDspp)
head(durEx)
#>                UID   PARAMETER RESULT
#> 8.N_PERENNIAL    8 N_PERENNIAL      4
#> 5.N_PERENNIAL    5 N_PERENNIAL      5
#> 9.N_PERENNIAL    9 N_PERENNIAL      7
#> 7.N_PERENNIAL    7 N_PERENNIAL      8
#> 10.N_PERENNIAL  10 N_PERENNIAL      9
#> 1.N_PERENNIAL    1 N_PERENNIAL     18
unique(durEx$PARAMETER)
#>  [1] "N_PERENNIAL"          "N_ANNUAL"             "N_ANN_PEREN"         
#>  [4] "N_ANN_BIEN"           "PCTN_PERENNIAL"       "PCTN_ANNUAL"         
#>  [7] "PCTN_ANN_PEREN"       "PCTN_ANN_BIEN"        "XABCOV_PERENNIAL"    
#> [10] "XABCOV_ANNUAL"        "XABCOV_ANN_PEREN"     "XABCOV_ANN_BIEN"     
#> [13] "XRCOV_PERENNIAL"      "XRCOV_ANNUAL"         "XRCOV_ANN_PEREN"     
#> [16] "XRCOV_ANN_BIEN"       "N_ANNUAL_NAT"         "PCTN_ANNUAL_NAT"     
#> [19] "XABCOV_ANNUAL_NAT"    "XRCOV_ANNUAL_NAT"     "N_ANNUAL_AC"         
#> [22] "PCTN_ANNUAL_AC"       "XABCOV_ANNUAL_AC"     "XRCOV_ANNUAL_AC"     
#> [25] "N_ANN_BIEN_NAT"       "PCTN_ANN_BIEN_NAT"    "XABCOV_ANN_BIEN_NAT" 
#> [28] "XRCOV_ANN_BIEN_NAT"   "N_ANN_BIEN_AC"        "PCTN_ANN_BIEN_AC"    
#> [31] "XABCOV_ANN_BIEN_AC"   "XRCOV_ANN_BIEN_AC"    "N_ANN_PEREN_NAT"     
#> [34] "PCTN_ANN_PEREN_NAT"   "XABCOV_ANN_PEREN_NAT" "XRCOV_ANN_PEREN_NAT" 
#> [37] "N_ANN_PEREN_AC"       "PCTN_ANN_PEREN_AC"    "XABCOV_ANN_PEREN_AC" 
#> [40] "XRCOV_ANN_PEREN_AC"   "N_PERENNIAL_NAT"      "PCTN_PERENNIAL_NAT"  
#> [43] "XABCOV_PERENNIAL_NAT" "XRCOV_PERENNIAL_NAT"  "N_PERENNIAL_AC"      
#> [46] "PCTN_PERENNIAL_AC"    "XABCOV_PERENNIAL_AC"  "XRCOV_PERENNIAL_AC"