Calculate growth habit metrics

This function calculates all growth habit metrics, with additional metrics if NWCA_NATSTAT variable is included in input data frame.

calcGrowthHabit(vascIn, sampID = "UID")

Arguments

vascIn

Data frame containing cover data summarized by UID and TAXON, with the following fields:

sampID: Variable(s) identified in sampID argument

TAXON: Taxon name

GROWTH_HABIT: USDA PLANTS GROWTH_HABIT variable with valid values FORB, GRAMINOID, HERB, SHRUB, SUBSHRUB, TREE, VINE, NONVASCULAR, combinations of these, or blank

GRH_ALT (optional): Combinations of GROWTH_HABIT variable as used in NWCA. Created in code if not supplied by user. Valid values include FORB, GRAMINOID, SHRUB, SSHRUB_FORB, SSHRUB_SHRUB, TREE, TREE_SHRUB, VINE, VINE_SHRUB.

TREE_COMB (optional): Indicator (1/0) for values of TREE or TREE_SHRUB in GRH_ALT. If not supplied by user, will be created in code

SHRUB_COMB (optional): Indicator value (0/1) for GRH_ALT values of SHRUB, SSHRUB_FORB, or SSHRUB_SHRUB. Created in code if not supplied by user.

HERB (optional): Indicator value (0/1) for GRH_ALT values of GRAMINOID or FORB. Created in code if not supplied by user.

XABCOV: Mean percent cover of taxon across plots

TOTN: Number of taxa in sample

sXRCOV: proportion of summed cover across all taxa (XTOTABCOV) represented by taxon in sample

NWCA_NATSTAT (optional): Native status variable with categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait value (represented as TRAITNM below):

N_TRAITNM: Number of taxa with trait

PCTN_TRAITNM: Number of taxa with trait as percentage of TOTN

XABCOV_TRAITNM: Sum of XABCOV values across taxa with trait

XRCOV_TRAITNM: Sum of sXRCOV values across taxa with trait

If NWCA_NATSTAT is in the input data frame, the same set except 'N_' metrics is calculated for native species and alien + cryptogenic species, with metric names suffixes of _NAT and _AC, respectively.

Details

Both GROWTH_HABIT and NWCA_NATSTAT variables are recoded to fewer categories. Taxa with 'UND' as native status are excluded.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom Blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

ghEx <- calcGrowthHabit(exPlant$byUIDspp)

head(ghEx)
#>           UID PARAMETER RESULT
#> 8.N_FORB    8    N_FORB      0
#> 1.N_FORB    1    N_FORB      1
#> 5.N_FORB    5    N_FORB      1
#> 9.N_FORB    9    N_FORB      4
#> 7.N_FORB    7    N_FORB      4
#> 10.N_FORB  10    N_FORB      6
unique(ghEx$PARAMETER)
#>   [1] "N_FORB"                "N_GRAMINOID"           "N_SHRUB"              
#>   [4] "N_TREE"                "N_TREE_SHRUB"          "N_VINE"               
#>   [7] "N_VINE_SHRUB"          "N_SSHRUB_FORB"         "N_SSHRUB_SHRUB"       
#>  [10] "PCTN_FORB"             "PCTN_GRAMINOID"        "PCTN_SHRUB"           
#>  [13] "PCTN_TREE"             "PCTN_TREE_SHRUB"       "PCTN_VINE"            
#>  [16] "PCTN_VINE_SHRUB"       "PCTN_SSHRUB_FORB"      "PCTN_SSHRUB_SHRUB"    
#>  [19] "XABCOV_FORB"           "XABCOV_GRAMINOID"      "XABCOV_SHRUB"         
#>  [22] "XABCOV_TREE"           "XABCOV_TREE_SHRUB"     "XABCOV_VINE"          
#>  [25] "XABCOV_VINE_SHRUB"     "XABCOV_SSHRUB_FORB"    "XABCOV_SSHRUB_SHRUB"  
#>  [28] "XRCOV_FORB"            "XRCOV_GRAMINOID"       "XRCOV_SHRUB"          
#>  [31] "XRCOV_TREE"            "XRCOV_TREE_SHRUB"      "XRCOV_VINE"           
#>  [34] "XRCOV_VINE_SHRUB"      "XRCOV_SSHRUB_FORB"     "XRCOV_SSHRUB_SHRUB"   
#>  [37] "N_TREE_COMB"           "PCTN_TREE_COMB"        "XABCOV_TREE_COMB"     
#>  [40] "XRCOV_TREE_COMB"       "N_SHRUB_COMB"          "PCTN_SHRUB_COMB"      
#>  [43] "XABCOV_SHRUB_COMB"     "XRCOV_SHRUB_COMB"      "N_HERB"               
#>  [46] "PCTN_HERB"             "XABCOV_HERB"           "XRCOV_HERB"           
#>  [49] "N_VINE_ALL"            "PCTN_VINE_ALL"         "XABCOV_VINE_ALL"      
#>  [52] "XRCOV_VINE_ALL"        "N_GRAMINOID_AC"        "PCTN_GRAMINOID_AC"    
#>  [55] "XABCOV_GRAMINOID_AC"   "XRCOV_GRAMINOID_AC"    "N_GRAMINOID_NAT"      
#>  [58] "XRCOV_GRAMINOID_NAT"   "XABCOV_GRAMINOID_NAT"  "PCTN_GRAMINOID_NAT"   
#>  [61] "N_FORB_AC"             "PCTN_FORB_AC"          "XABCOV_FORB_AC"       
#>  [64] "XRCOV_FORB_AC"         "N_FORB_NAT"            "XRCOV_FORB_NAT"       
#>  [67] "XABCOV_FORB_NAT"       "PCTN_FORB_NAT"         "N_HERB_AC"            
#>  [70] "PCTN_HERB_AC"          "XABCOV_HERB_AC"        "XRCOV_HERB_AC"        
#>  [73] "N_HERB_NAT"            "PCTN_HERB_NAT"         "XABCOV_HERB_NAT"      
#>  [76] "XRCOV_HERB_NAT"        "N_SHRUB_COMB_AC"       "PCTN_SHRUB_COMB_AC"   
#>  [79] "XABCOV_SHRUB_COMB_AC"  "XRCOV_SHRUB_COMB_AC"   "N_SHRUB_COMB_NAT"     
#>  [82] "PCTN_SHRUB_COMB_NAT"   "XABCOV_SHRUB_COMB_NAT" "XRCOV_SHRUB_COMB_NAT" 
#>  [85] "N_TREE_COMB_AC"        "PCTN_TREE_COMB_AC"     "XABCOV_TREE_COMB_AC"  
#>  [88] "XRCOV_TREE_COMB_AC"    "N_TREE_COMB_NAT"       "PCTN_TREE_COMB_NAT"   
#>  [91] "XABCOV_TREE_COMB_NAT"  "XRCOV_TREE_COMB_NAT"   "N_VINE_AC"            
#>  [94] "PCTN_VINE_AC"          "XABCOV_VINE_AC"        "XRCOV_VINE_AC"        
#>  [97] "N_VINE_NAT"            "PCTN_VINE_NAT"         "XABCOV_VINE_NAT"      
#> [100] "XRCOV_VINE_NAT"        "N_VINE_SHRUB_AC"       "PCTN_VINE_SHRUB_AC"   
#> [103] "XABCOV_VINE_SHRUB_AC"  "XRCOV_VINE_SHRUB_AC"   "N_VINE_SHRUB_NAT"     
#> [106] "PCTN_VINE_SHRUB_NAT"   "XABCOV_VINE_SHRUB_NAT" "XRCOV_VINE_SHRUB_NAT" 
#> [109] "N_VINE_ALL_AC"         "PCTN_VINE_ALL_AC"      "XABCOV_VINE_ALL_AC"   
#> [112] "XRCOV_VINE_ALL_AC"     "N_VINE_ALL_NAT"        "PCTN_VINE_ALL_NAT"    
#> [115] "XABCOV_VINE_ALL_NAT"   "XRCOV_VINE_ALL_NAT"