Calculate richness metrics

This function calculates richness metrics using plot- and UID-based datasets at the species, genus, and family levels.

calcRichness(
  byUIDspp,
  byPlotspp,
  byUIDgen,
  byPlotgen,
  byUIDfam,
  byPlotfam,
  sampID = "UID"
)

Arguments

byUIDspp

Data frame containing species data summarized to the UID and TAXON level, with the following variables:

sampID: Variables identified by sampID argument

TAXON: Taxon name

COVER: Total percent cover for taxon

DISTINCT: Distinctness (0/1) of taxon within sample, where taxa above species level are not distinct (0) if taxa at a lower level are also included in sample.

byPlotspp

Data frame containing species data summarized to the UID, PLOT, and TAXON level, with the following variables:

sampID: Variables identified in sampID argument

PLOT: Plot from which data were collected

TAXON: Taxon name

COVER: Total percent cover for taxon

DISTINCT: Distinctness (0/1) of taxon within sample, where taxa above species level are not distinct (0) if taxa at a lower level are also included in sample.

byUIDgen

Data frame containing genus-level data summarized to the sampID variables and TAXON level and containing the same variables as byUIDspp (described above).

byPlotgen

Data frame containing genus-level data summarized to the sampID variables, PLOT, and TAXON level and containing the same variables as byPlotspp (described above).

byUIDfam

Data frame containing family-level data summarized to the sampID variables and TAXON level and containing the same variables as byUIDspp (described above).

byPlotfam

Data frame containing family-level data summarized to the sampID variables, PLOT, and TAXON level and containing the same variables as byPlotspp (described above).

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing sampID variables, PARAMETER, and RESULT, with one row of results per parameter and sampID. The values for PARAMETER consist of the metric name concatenated with taxonomic level (represented as SPP, GEN, and FAM below):

TOTN_SPP, TOTN_GEN, TOTN_FAM: Number of unique taxa in sample (UID)

XN_SPP, XN_GEN, XN_FAM: Mean number of taxa per plot

MEDN_SPP, MEDN_GEN, MEDN_FAM: Median number of taxa per plot

SDN_SPP, SDN_GEN, SDN_FAM: Standard deviation of number of taxa per plot

N_PLOTS: Number of plots sampled for UID

If NWCA_NATSTAT is present in the input data frame, the following metrics are calculated. GRP in the metric names below represents native status values of NAT, ADV, CRYP, INTR, and subsets ALIEN (INTR + ADV) and AC (ALIEN and CRYP):

TOTN_GRP: Number of unique taxa in sample (UID)

XN_GRP: Mean number of taxa per plot

MEDN_GRP: Median number of taxa per plot

SDN_GRP: Standard deviation of number of taxa per plot

Details

The prepareData() function creates a list object with all of the necessary input data frames. For each taxonomic level, the function createDFs() creates a list with a data frame summarized by UID and one by sampID variables and PLOT.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

richEx <- calcRichness(
  exPlant$byUIDspp, exPlant$byPlotspp,
  exPlant$byUIDgen, exPlant$byPlotgen, exPlant$byUIDfam, exPlant$byPlotfam
)

head(richEx)
#>             UID PARAMETER RESULT
#> 1.TOTN_SPP    1  TOTN_SPP     18
#> 10.TOTN_SPP  10  TOTN_SPP     10
#> 2.TOTN_SPP    2  TOTN_SPP     49
#> 3.TOTN_SPP    3  TOTN_SPP     21
#> 4.TOTN_SPP    4  TOTN_SPP     34
#> 5.TOTN_SPP    5  TOTN_SPP      6
unique(richEx$PARAMETER)
#>  [1] "TOTN_SPP"      "XN_SPP"        "MEDN_SPP"      "SDN_SPP"      
#>  [5] "TOTN_GEN"      "XN_GEN"        "MEDN_GEN"      "SDN_GEN"      
#>  [9] "TOTN_FAM"      "XN_FAM"        "MEDN_FAM"      "SDN_FAM"      
#> [13] "TOTN_NATSPP"   "XN_NATSPP"     "MEDN_NATSPP"   "SDN_NATSPP"   
#> [17] "TOTN_ADVSPP"   "XN_ADVSPP"     "MEDN_ADVSPP"   "SDN_ADVSPP"   
#> [21] "TOTN_CRYPSPP"  "XN_CRYPSPP"    "MEDN_CRYPSPP"  "SDN_CRYPSPP"  
#> [25] "TOTN_INTRSPP"  "XN_INTRSPP"    "MEDN_INTRSPP"  "SDN_INTRSPP"  
#> [29] "TOTN_ALIENSPP" "XN_ALIENSPP"   "MEDN_ALIENSPP" "SDN_ALIENSPP" 
#> [33] "TOTN_AC"       "XN_AC"         "MEDN_AC"       "SDN_AC"