Calculate metrics based on S&T (Status and Trends) categories

This function calculations metrics based on the number of S&T classes found across plots. This function is called by calcVtype_GcovMets().

calcSandTMets(dataIn, nPlot, sampID = "UID")

Arguments

dataIn

A data frame containing the following variables:

sampID: Variables identified in sampID argument

PLOT: Sample plot from which data were collected

PARAMETER: specific measurement type

RESULT: measured value

The following parameters are used in calculating wetland type metrics: 'SANDT_CLASS' and 'PAL_FARMED', or 'WETLAND_TYPE' (2016). Additional parameters or variables are ignored.

nPlot

A data frame with the number of plots sampled associated with each sample, with sampID variables and NPLOTS

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default.

Value

Either a character string containing an error message when metric calculation is not successful, or a data frame. Data frame contains sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait value, with valid values of: N_SANDT, DOM_SANDT, D_SANDT, H_SANDT, J_SANDT. A list of metric descriptions is provided in the document named VegTypes_GrdCover_Metric_Descriptions.pdf included in the help directory for the package.

Details

If any of the parameters are missing, they are assumed to be zeros (if numeric), and metric values associated with any metrics that cannot be calculated due to missing parameters are set to a standardized value.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(Vtype_GrCovEx)
#>   UID PLOT       PARAMETER RESULT
#> 1   7    4   MINIMUM_DEPTH      8
#> 2   7    3 LITTER_DEPTH_NE      0
#> 3   7    5        TALL_VEG      0
#> 4   7    4     WATER_AQVEG      0
#> 5   7    5   MINIMUM_DEPTH      8
#> 6   7    5     WATER_NOVEG      0
# Create data frame with number of plots sampled for each sampID
nplots <- data.frame(UID = seq(1:10), NPLOTS = rep(5, 10), stringsAsFactors = F)
# alternative approach to creating this data frame

sandtEx <- calcSandTMets(Vtype_GrCovEx, nplots)
#> [1] "Done with veg heterogeneity metrics"

head(sandtEx)
#>   UID PARAMETER RESULT
#> 1   1   D_SANDT      0
#> 2   1   N_SANDT      1
#> 3   1   H_SANDT      0
#> 4   1   J_SANDT      0
#> 5   1 DOM_SANDT    PFO
#> 6   2 DOM_SANDT    PFO
unique(sandtEx$PARAMETER)
#> [1] "D_SANDT"   "N_SANDT"   "H_SANDT"   "J_SANDT"   "DOM_SANDT"