Calculate tree cover metrics

This function calculates tree cover metrics using tree data, and is called by function calcTreeMetrics().

calcTreeCoverMets(treeIn, nPlot, sampID = "UID")

Arguments

treeIn

A data frame containing the following variables:

sampID: Variable(s) identified in sampID argument

PAGE: Page from field form

LINE: Line number from field form

PLOT: Sample plot from which data were collected

PARAMETER: specific measurement type

RESULT: measured value

The following parameters are used in calculating tree metrics: 'VSMALL_TREE', 'SMALL_TREE', 'LMED_TREE' , 'HMED_TREE', 'TALL_TREE', 'VTALL_TREE'. Additional parameters or variables are ignored.

nPlot

A data frame with the number of plots sampled associated with each sample, with sampID variables and NPLOTS.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Either a character string containing an error message when metric calculation is not successful, or a data frame. The data frame contains the sampID variables, PARAMETER, RESULT, where PARAMETER values include: N_TREESPP, N_TALL_TREE, N_HMED_TREE, N_LMED_TREE, N_SMALL_TREE, N_VSMALL_TREE, N_VTALL_TREE, N_TREE_UPPER, N_TREE_MID, N_TREE_GROUND, PCTN_TREE_UPPER, PCTN_TREE_MID, PCTN_TREE_GROUND. A list of metric descriptions is provided in the document named Tree_Metric_Descriptions.pdf included in the help directory for the package.

Details

If any of the parameters are missing, they are assumed to be zeros, and metric values associated with any metrics that cannot be calculated due to missing parameters are set to 0.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(TreesEx)
#>   UID PLOT PAGE LINE       PARAMETER RESULT
#> 1   7    1    1    2 SPECIES_NAME_ID  36494
#> 2   7    1    1    2       LMED_TREE      5
#> 3   7    5    1    3       LMED_TREE      0
#> 4   7    5    1    3      SMALL_TREE      1
#> 5   7    1    1    2       TALL_TREE      0
#> 6   7    1    1    2         JR_TREE      0
nplots <- aggregate(
  x = list(NPLOTS = TreesEx$PLOT), by = TreesEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

tcvrEx <- calcTreeCoverMets(TreesEx, nPlot = nplots, sampID = "UID")
#> [1] "Done with tree species metrics"
#> [1] "Done with tree cover metrics"

head(tcvrEx)
#>   UID RESULT       PARAMETER
#> 1   1      3       N_TREESPP
#> 2   1      0  IMP_VTALL_TREE
#> 3   1      0 IMP_VSMALL_TREE
#> 4   1     80  FREQ_TALL_TREE
#> 5   1      0 XCOV_VTALL_TREE
#> 6   1      0   IMP_LMED_TREE
unique(tcvrEx$PARAMETER)
#>  [1] "N_TREESPP"        "IMP_VTALL_TREE"   "IMP_VSMALL_TREE"  "FREQ_TALL_TREE"  
#>  [5] "XCOV_VTALL_TREE"  "IMP_LMED_TREE"    "FREQ_SMALL_TREE"  "FREQ_VTALL_TREE" 
#>  [9] "IMP_TALL_TREE"    "FREQ_HMED_TREE"   "N_HMED_TREE"      "N_VSMALL_TREE"   
#> [13] "FREQ_TREE_GROUND" "FREQ_LMED_TREE"   "N_TALL_TREE"      "PCTN_TREE_UPPER" 
#> [17] "XCOV_TREE_GROUND" "N_SMALL_TREE"     "PCTN_TREE_MID"    "FREQ_TREE_UPPER" 
#> [21] "N_LMED_TREE"      "PCTN_TREE_GROUND" "FREQ_TREE_MID"    "FREQ_VSMALL_TREE"
#> [25] "XCOV_LMED_TREE"   "N_TREE_GROUND"    "XCOV_TREE_MID"    "XCOV_TALL_TREE"  
#> [29] "IMP_SMALL_TREE"   "N_VTALL_TREE"     "XCOV_SMALL_TREE"  "IMP_HMED_TREE"   
#> [33] "N_TREE_UPPER"     "XCOV_HMED_TREE"   "XCOV_VSMALL_TREE" "N_TREE_MID"      
#> [37] "XCOV_TREE_UPPER"  "IMP_TREE_GROUND"  "IMP_TREE_MID"     "IMP_TREE_UPPER"