Calculate NWCA 2011 tree metrics based on data collected from field form V-4a

This function calculates the NWCA 2011 tree metrics. It assumes input data are organized in long format so that each row represents the value for a single parameter value within a single plot at a site.

calcTreeMets(treeIn, nPlotIn, sampID = "UID")

Arguments

treeIn

A data frame containing the following variables:

sampID: Variable(s) found in the argument sampID

PAGE: Page number from field form

LINE: Line number from field form

PLOT: Sample plot from which data were collected

PARAMETER: specific measurement type

RESULT: measured value

The following parameters are used in calculating tree metrics: 'XXTHIN_SNAG','XTHIN_SNAG','THIN_SNAG', 'JR_SNAG','THICK_SNAG','XTHICK_SNAG','XXTHICK_SNAG','XXTHIN_TREE', 'XTHIN_TREE','THIN_TREE','JR_TREE','THICK_TREE','XTHICK_TREE', 'XXTHICK_TREE','VSMALL_TREE','SMALL_TREE','LMED_TREE','HMED_TREE', 'TALL_TREE','VTALL_TREE'. Additional parameters or variables are ignored.

nPlotIn

A data frame with the number of plots sampled associated with each sample, with sampID variables and NPLOTS.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Either a character string containing an error message when metric calculation is not successful, or a data frame. The first columns of the data frame are the sampID variables and subsequent columns are named for each metric and contain metric values. A list of tree metrics is installed with the package. Use system.file("Tree_Metric_Descriptions.pdf", package="aquametNWCA") to locate the file.

Details

If any of the parameters are missing, they are assumed to be zeros, and metric values associated with any metrics that cannot be calculated due to missing parameters are set to 0.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(TreesEx)
#>   UID PLOT PAGE LINE       PARAMETER RESULT
#> 1   7    1    1    2 SPECIES_NAME_ID  36494
#> 2   7    1    1    2       LMED_TREE      5
#> 3   7    5    1    3       LMED_TREE      0
#> 4   7    5    1    3      SMALL_TREE      1
#> 5   7    1    1    2       TALL_TREE      0
#> 6   7    1    1    2         JR_TREE      0

nplots <- aggregate(
  x = list(NPLOTS = TreesEx$PLOT), by = TreesEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

exOut <- calcTreeMets(TreesEx, nPlotIn = nplots, sampID = "UID")
#> [1] "Done with snag metrics"
#> [1] "Done with tree count metrics"
#> [1] "Done with tree species metrics"
#> [1] "Done with tree cover metrics"

head(exOut)
#>    UID TOTN_SNAGS TOTN_THIN_SNAG XN_JR_SNAG XN_XTHIN_SNAG TOTN_THICK_SNAG
#> 1    1          0              0          0           0.0               0
#> 15   2          2              1          0           0.2               0
#> 29   3          0              0          0           0.0               0
#> 43   4          0              0          0           0.0               0
#> 57   5          0              0          0           0.0               0
#> 71   6          2              1          0           0.2               0
#>    TOTN_XXTHIN_SNAG XN_XTHICK_SNAG TOTN_JR_SNAG TOTN_XTHIN_SNAG XN_THIN_SNAG
#> 1                 0              0            0               0          0.0
#> 15                0              0            0               1          0.2
#> 29                0              0            0               0          0.0
#> 43                0              0            0               0          0.0
#> 57                0              0            0               0          0.0
#> 71                0              0            0               1          0.2
#>    XN_SNAGS TOTN_XTHICK_SNAG XN_THICK_SNAG XN_XXTHIN_SNAG TOTN_TREES
#> 1       0.0                0             0              0         24
#> 15      0.4                0             0              0         28
#> 29      0.0                0             0              0          0
#> 43      0.0                0             0              0         79
#> 57      0.0                0             0              0          0
#> 71      0.4                0             0              0         45
#>    TOTN_THIN_TREE TOTN_XXTHIN_TREE XN_XTHICK_TREE TOTN_THICK_TREE
#> 1               6                2              0               0
#> 15              9                1              0               0
#> 29              0                0              0               0
#> 43              6               44              0               1
#> 57              0                0              0               0
#> 71              5               24              0               0
#>    TOTN_XXTHICK_TREE XN_THIN_TREE TOTN_JR_TREE TOTN_XTHIN_TREE XN_THICK_TREE
#> 1                  0          1.2            3              13           0.0
#> 15                 0          1.8            6              12           0.0
#> 29                 0          0.0            0               0           0.0
#> 43                 0          1.2           11              17           0.2
#> 57                 0          0.0            0               0           0.0
#> 71                 0          1.0            0              16           0.0
#>    XN_TREES TOTN_XTHICK_TREE XN_JR_TREE XN_XTHIN_TREE TOTN_MID XN_LARGE
#> 1       4.8                0        0.6           2.6        9      0.0
#> 15      5.6                0        1.2           2.4       15      0.0
#> 29      0.0                0        0.0           0.0        0      0.0
#> 43     15.8                0        2.2           3.4       17      0.2
#> 57      0.0                0        0.0           0.0        0      0.0
#> 71      9.0                0        0.0           3.2        5      0.0
#>    XN_XXTHIN_TREE TOTN_SMALL XN_MID XN_XXTHICK_TREE TOTN_LARGE XN_SMALL
#> 1             0.4         15    1.8               0          0      3.0
#> 15            0.2         13    3.0               0          0      2.6
#> 29            0.0          0    0.0               0          0      0.0
#> 43            8.8         61    3.4               0          1     12.2
#> 57            0.0          0    0.0               0          0      0.0
#> 71            4.8         40    1.0               0          0      8.0
#>    N_TREESPP IMP_VTALL_TREE IMP_VSMALL_TREE FREQ_TALL_TREE XCOV_VTALL_TREE
#> 1          3              0             0.0             80               0
#> 15         5              0             0.0            100               0
#> 29         0              0             0.0              0               0
#> 43        12              0            35.1             80               0
#> 57         0              0             0.0              0               0
#> 71         4              0             0.0             80               0
#>    IMP_LMED_TREE FREQ_SMALL_TREE FREQ_VTALL_TREE IMP_TALL_TREE FREQ_HMED_TREE
#> 1            0.0              40               0          92.0             40
#> 15           0.0               0               0          64.4             80
#> 29           0.0               0               0           0.0              0
#> 43          58.7             100               0          56.9            100
#> 57           0.0               0               0           0.0              0
#> 71          11.0               0               0          48.0            100
#>    N_HMED_TREE N_VSMALL_TREE FREQ_TREE_GROUND FREQ_LMED_TREE N_TALL_TREE
#> 1            3             0               40              0           3
#> 15           3             0                0              0           3
#> 29           0             0                0              0           0
#> 43           7             6              100            100           6
#> 57           0             0                0              0           0
#> 71           4             0                0             20           1
#>    PCTN_TREE_UPPER XCOV_TREE_GROUND N_SMALL_TREE PCTN_TREE_MID FREQ_TREE_UPPER
#> 1              100              1.4            2           100              80
#> 15              60              0.0            0            60             100
#> 29               0              0.0            0             0               0
#> 43              50             13.2            5            75              80
#> 57               0              0.0            0             0               0
#> 71              25              0.0            0           100              80
#>    N_LMED_TREE PCTN_TREE_GROUND FREQ_TREE_MID FREQ_VSMALL_TREE XCOV_LMED_TREE
#> 1            0            66.67            40                0            0.0
#> 15           0             0.00            80                0            0.0
#> 29           0             0.00             0                0            0.0
#> 43           8            58.33           100               60           17.4
#> 57           0             0.00             0                0            0.0
#> 71           1             0.00           100                0            2.0
#>    N_TREE_GROUND XCOV_TREE_MID XCOV_TALL_TREE IMP_SMALL_TREE N_VTALL_TREE
#> 1              2           6.0          104.0           20.7            0
#> 15             0           3.2           28.8            0.0            0
#> 29             0           0.0            0.0            0.0            0
#> 43             7          44.4           33.8           51.5            0
#> 57             0           0.0            0.0            0.0            0
#> 71             0          20.8           16.0            0.0            0
#>    XCOV_SMALL_TREE IMP_HMED_TREE N_TREE_UPPER XCOV_HMED_TREE XCOV_VSMALL_TREE
#> 1              1.4          23.0            3            6.0              0.0
#> 15             0.0          41.6            3            3.2              0.0
#> 29             0.0           0.0            0            0.0              0.0
#> 43             3.0          63.5            6           27.0             10.2
#> 57             0.0           0.0            0            0.0              0.0
#> 71             0.0          59.4            1           18.8              0.0
#>    N_TREE_MID XCOV_TREE_UPPER IMP_TREE_GROUND IMP_TREE_MID IMP_TREE_UPPER
#> 1           3           104.0            20.7         23.0           92.0
#> 15          3            28.8             0.0         41.6           64.4
#> 29          0             0.0             0.0          0.0            0.0
#> 43          9            33.8            56.6         72.2           56.9
#> 57          0             0.0             0.0          0.0            0.0
#> 71          4            16.0             0.0         60.4           48.0
str(exOut)
#> 'data.frame':	10 obs. of  77 variables:
#>  $ UID              : int  1 2 3 4 5 6 7 8 9 10
#>  $ TOTN_SNAGS       : num  0 2 0 0 0 2 0 0 0 0
#>  $ TOTN_THIN_SNAG   : num  0 1 0 0 0 1 0 0 0 0
#>  $ XN_JR_SNAG       : num  0 0 0 0 0 0 0 0 0 0
#>  $ XN_XTHIN_SNAG    : num  0 0.2 0 0 0 0.2 0 0 0 0
#>  $ TOTN_THICK_SNAG  : num  0 0 0 0 0 0 0 0 0 0
#>  $ TOTN_XXTHIN_SNAG : num  0 0 0 0 0 0 0 0 0 0
#>  $ XN_XTHICK_SNAG   : num  0 0 0 0 0 0 0 0 0 0
#>  $ TOTN_JR_SNAG     : num  0 0 0 0 0 0 0 0 0 0
#>  $ TOTN_XTHIN_SNAG  : num  0 1 0 0 0 1 0 0 0 0
#>  $ XN_THIN_SNAG     : num  0 0.2 0 0 0 0.2 0 0 0 0
#>  $ XN_SNAGS         : num  0 0.4 0 0 0 0.4 0 0 0 0
#>  $ TOTN_XTHICK_SNAG : num  0 0 0 0 0 0 0 0 0 0
#>  $ XN_THICK_SNAG    : num  0 0 0 0 0 0 0 0 0 0
#>  $ XN_XXTHIN_SNAG   : num  0 0 0 0 0 0 0 0 0 0
#>  $ TOTN_TREES       : num  24 28 0 79 0 45 0 0 0 0
#>  $ TOTN_THIN_TREE   : num  6 9 0 6 0 5 0 0 0 0
#>  $ TOTN_XXTHIN_TREE : num  2 1 0 44 0 24 0 0 0 0
#>  $ XN_XTHICK_TREE   : num  0 0 0 0 0 0 0 0 0 0
#>  $ TOTN_THICK_TREE  : num  0 0 0 1 0 0 0 0 0 0
#>  $ TOTN_XXTHICK_TREE: num  0 0 0 0 0 0 0 0 0 0
#>  $ XN_THIN_TREE     : num  1.2 1.8 0 1.2 0 1 0 0 0 0
#>  $ TOTN_JR_TREE     : num  3 6 0 11 0 0 0 0 0 0
#>  $ TOTN_XTHIN_TREE  : num  13 12 0 17 0 16 0 0 0 0
#>  $ XN_THICK_TREE    : num  0 0 0 0.2 0 0 0 0 0 0
#>  $ XN_TREES         : num  4.8 5.6 0 15.8 0 9 0 0 0 0
#>  $ TOTN_XTHICK_TREE : num  0 0 0 0 0 0 0 0 0 0
#>  $ XN_JR_TREE       : num  0.6 1.2 0 2.2 0 0 0 0 0 0
#>  $ XN_XTHIN_TREE    : num  2.6 2.4 0 3.4 0 3.2 0 0 0 0
#>  $ TOTN_MID         : num  9 15 0 17 0 5 0 0 0 0
#>  $ XN_LARGE         : num  0 0 0 0.2 0 0 0 0 0 0
#>  $ XN_XXTHIN_TREE   : num  0.4 0.2 0 8.8 0 4.8 0 0 0 0
#>  $ TOTN_SMALL       : num  15 13 0 61 0 40 0 0 0 0
#>  $ XN_MID           : num  1.8 3 0 3.4 0 1 0 0 0 0
#>  $ XN_XXTHICK_TREE  : num  0 0 0 0 0 0 0 0 0 0
#>  $ TOTN_LARGE       : num  0 0 0 1 0 0 0 0 0 0
#>  $ XN_SMALL         : num  3 2.6 0 12.2 0 8 0 0 0 0
#>  $ N_TREESPP        : num  3 5 0 12 0 4 1 0 0 0
#>  $ IMP_VTALL_TREE   : num  0 0 0 0 0 0 0 0 0 0
#>  $ IMP_VSMALL_TREE  : num  0 0 0 35.1 0 0 0 0 0 0
#>  $ FREQ_TALL_TREE   : num  80 100 0 80 0 80 0 0 0 0
#>  $ XCOV_VTALL_TREE  : num  0 0 0 0 0 0 0 0 0 0
#>  $ IMP_LMED_TREE    : num  0 0 0 58.7 0 11 21.1 0 0 0
#>  $ FREQ_SMALL_TREE  : num  40 0 0 100 0 0 60 0 0 0
#>  $ FREQ_VTALL_TREE  : num  0 0 0 0 0 0 0 0 0 0
#>  $ IMP_TALL_TREE    : num  92 64.4 0 56.9 0 48 0 0 0 0
#>  $ FREQ_HMED_TREE   : num  40 80 0 100 0 100 0 0 0 0
#>  $ N_HMED_TREE      : num  3 3 0 7 0 4 0 0 0 0
#>  $ N_VSMALL_TREE    : num  0 0 0 6 0 0 0 0 0 0
#>  $ FREQ_TREE_GROUND : num  40 0 0 100 0 0 60 0 0 0
#>  $ FREQ_LMED_TREE   : num  0 0 0 100 0 20 40 0 0 0
#>  $ N_TALL_TREE      : num  3 3 0 6 0 1 0 0 0 0
#>  $ PCTN_TREE_UPPER  : num  100 60 0 50 0 25 0 0 0 0
#>  $ XCOV_TREE_GROUND : num  1.4 0 0 13.2 0 0 1.2 0 0 0
#>  $ N_SMALL_TREE     : num  2 0 0 5 0 0 1 0 0 0
#>  $ PCTN_TREE_MID    : num  100 60 0 75 0 100 100 0 0 0
#>  $ FREQ_TREE_UPPER  : num  80 100 0 80 0 80 0 0 0 0
#>  $ N_LMED_TREE      : num  0 0 0 8 0 1 1 0 0 0
#>  $ PCTN_TREE_GROUND : num  66.7 0 0 58.3 0 ...
#>  $ FREQ_TREE_MID    : num  40 80 0 100 0 100 40 0 0 0
#>  $ FREQ_VSMALL_TREE : num  0 0 0 60 0 0 0 0 0 0
#>  $ XCOV_LMED_TREE   : num  0 0 0 17.4 0 2 2.2 0 0 0
#>  $ N_TREE_GROUND    : num  2 0 0 7 0 0 1 0 0 0
#>  $ XCOV_TREE_MID    : num  6 3.2 0 44.4 0 20.8 2.2 0 0 0
#>  $ XCOV_TALL_TREE   : num  104 28.8 0 33.8 0 16 0 0 0 0
#>  $ IMP_SMALL_TREE   : num  20.7 0 0 51.5 0 0 30.6 0 0 0
#>  $ N_VTALL_TREE     : num  0 0 0 0 0 0 0 0 0 0
#>  $ XCOV_SMALL_TREE  : num  1.4 0 0 3 0 0 1.2 0 0 0
#>  $ IMP_HMED_TREE    : num  23 41.6 0 63.5 0 59.4 0 0 0 0
#>  $ N_TREE_UPPER     : num  3 3 0 6 0 1 0 0 0 0
#>  $ XCOV_HMED_TREE   : num  6 3.2 0 27 0 18.8 0 0 0 0
#>  $ XCOV_VSMALL_TREE : num  0 0 0 10.2 0 0 0 0 0 0
#>  $ N_TREE_MID       : num  3 3 0 9 0 4 1 0 0 0
#>  $ XCOV_TREE_UPPER  : num  104 28.8 0 33.8 0 16 0 0 0 0
#>  $ IMP_TREE_GROUND  : num  20.7 0 0 56.6 0 0 30.6 0 0 0
#>  $ IMP_TREE_MID     : num  23 41.6 0 72.2 0 60.4 21.1 0 0 0
#>  $ IMP_TREE_UPPER   : num  92 64.4 0 56.9 0 48 0 0 0 0
#>  - attr(*, "reshapeWide")=List of 5
#>   ..$ v.names: chr "RESULT"
#>   ..$ timevar: chr "PARAMETER"
#>   ..$ idvar  : chr "UID"
#>   ..$ times  : chr [1:76] "TOTN_SNAGS" "TOTN_THIN_SNAG" "XN_JR_SNAG" "XN_XTHIN_SNAG" ...
#>   ..$ varying: chr [1, 1:76] "RESULT.TOTN_SNAGS" "RESULT.TOTN_THIN_SNAG" "RESULT.XN_JR_SNAG" "RESULT.XN_XTHIN_SNAG" ...