Calculate 2011 Vegetation MMI and assign condition class

This function calculates the NWCA 2011 Vegetation Multimetric Index (VMMI) from metric inputs. If the appropriate variable(s) describing site ecoregion and wetland group type according to NWCA 2011 are included in the input data frame, condition class (Good/Fair/Poor) will also be assigned.

calcVMMI_fromMets(metsIn, sampID = "UID")

Arguments

metsIn

Data frame containing, at a minimum:

sampID: Variables identified in sampID argument

FQAI_ALL: Floristic Quality Assessment Index based on all species

N_TOL: Number of tolerant species

RIMP_NATSPP: Relative importance of native species

XRCOV_MONOCOTS_NAT: Mean relative cover of native monocots

Optional:

NWCA_ECO4: Aggregated ecoregion abbreviations as used in NWCA 2011. Valid values include: CPL (Coastal Plains), EMU (Eastern Mountains and Uplands), IPL (Interior Plains), and W (West).

NWCA_WET_GRP: Wetland type groups as used in NWCA 2011. Valid values include: EH (Estuarine Herbaceous), EW (Estuarine Woody), PRLH (Palustrine, Riverine, and Lacustrine Herbaceous), PRLW (Palustrine, Riverine, and Lacustrine Woody).

ECO_X_WETGRP: Combinations of NWCA_ECO4 and NWCA_WET_GRP. Valid values include ALL-EH, ALL-EW, CPL-PRLH, CPL-PRLW, EMU-PRLH, EMU-PRLW, IPL-PRLH, IPL-PRLW, W-PRLH, W-PRLW

To assign condition class, either NWCA_ECO4 and NWCA_WET_GRP need to be included OR ECO_X_WETGRP needs to be included in the input data frame.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing:

sampID Variable(s) found in the argument sampID

FQAI_ALL_SC: Scored FQAI_ALL, on 0-10 scale

N_TOL_SC: Scored N_TOL, on 0-10 scale

RIMP_NATSPP_SC: Scored RIMP_NATSPP, on 0-10 scale

XRCOV_MONOCOTS_NAT_SC: Scored XRCOV_MONOCOTS_NAT, on 0-10 scale

VMMI: NWCA 2011 Vegetation Multimetric Index on 100-point scale

If ECO_X_WETGRP is provided or can be determined from NWCA_ECO4 and NWCA_WET_GRP, ECO_X_WETGRP and the following is also provided as output:

VEGCOND: Vegetation condition class, based on VMMI score and site location, as assigned for NWCA 2011. Valid values are Good, Fair, and Poor.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(vmmiMetEx)
#>   UID NWCA_ECO4 NWCA_WET_GRP ECO_X_WETGRP FQAI_ALL N_TOL RIMP_NATSPP
#> 1   1       CPL         PRLW     CPL-PRLW    20.27     8      100.00
#> 2   2       CPL         PRLW     CPL-PRLW    31.43    16       96.29
#> 3   3       IPL         PRLH     IPL-PRLH    21.39     9       87.97
#> 4   4       CPL         PRLW     CPL-PRLW    30.41    11       96.73
#> 5   5       CPL         PRLH     CPL-PRLH    16.33     0      100.00
#> 6   6       IPL         PRLW     IPL-PRLW    40.23    36       95.42
#>   XRCOV_MONOCOTS_NAT
#> 1               0.60
#> 2               3.73
#> 3              89.47
#> 4               3.27
#> 5              98.37
#> 6              19.88

# Test using ECO_X_WETGRP variable in data frame
check.1 <- calcVMMI_fromMets(vmmiMetEx)

# Now drop ECO_X_WETGRP and use NWCA_ECO4 and NWCA_WET_GRP
check.2 <- calcVMMI_fromMets(subset(vmmiMetEx, select = -ECO_X_WETGRP))

# Now run without ECO_X_WETGRP, NWCA_ECO4, and NWCA_WET_GRP
check.3 <- calcVMMI_fromMets(subset(vmmiMetEx, select = c(
  "UID", "FQAI_ALL",
  "N_TOL", "RIMP_NATSPP", "XRCOV_MONOCOTS_NAT"
)))
#> [1] "Warning: Must include variables NWCA_WET_GRP and NWCA_ECO4 OR ECO_X_WETGRP to determine condition class. This variable is missing, and condition class will not be assigned, but VMMI will be calculated."