Calculate NWCA 2011 vascular plant metrics based on form V-2

This function calculates the NWCA 2011 vascular plant metrics. It assumes input data are organized so that each row represents the cover for a single taxon within a single plot at a site. It also assumes a single column is used to identify each unique sample.

calcVascPlantMets(
  vascIn,
  taxon_name,
  taxaIn = taxaNWCA,
  taxaNat = ccNatNWCA,
  taxaCC = ccNatNWCA,
  taxaWIS = wisNWCA,
  sampID = "UID",
  state = "STATE",
  coeReg = "USAC_REGION",
  cValReg = "NWC_CREG"
)

Arguments

vascIn

A data frame with the following variables (at a minimum):

sampID: Variable(s) identified in sampID argument, 'UID' by default

PLOT: plot number

Variable named in state: Two-letter state postal code of site, used to link native status to taxa in native status taxalist (inNat)

Variable named in coeReg: U.S. Army Corps of Engineers region abbreviation for sample, to correspond to GEOG_ID in Wetland Indicator Status taxalist (inWIS)

Variable named in cValReg: NWCA C-value regions: values must match GEOG_ID in C-value taxalist (inCVal)

Variable named in taxon_name: Taxon name, must match with taxa data frame

COVER: value from 0 to 100 indicating relative cover of taxon within plot). Data should already be summed by UID, PLOT, and USDA_NAME.

taxon_name

String containing the name of variable for taxon name in

taxaIn

A data frame containing taxonomy of all taxa found in vascIn, with the following variables at a minimum:

USDA_NAME: USDA accepted name of taxon in sample

FAMILY: family name of taxon

GENUS: genus name of taxon

CATEGORY: USDA PLANTS category

GROWTH_HABIT: growth habit as designated in USDA PLANTS

DURATION: duration as designated by USDA PLANTS

If this data frame is not supplied, the data set taxaNWCA included in the package is used.

taxaNat

A data frame containing native status as assigned for NWCA for all taxa in vascIn, with the following variables, at a minimum:

USDA_NAME: USDA accepted name

GEOG_ID: state to which value applies

NWCA_NATSTAT: state- and taxon-specific native status

If this data frame is not supplied, the data set ccNWCA included in the package is used.

taxaCC

A data frame containing Coefficient of Conservatism as assigned for NWCA for all taxa in vascIn, with the following variables, at a minimum:

USDA_NAME: USDA accepted name

GEOG_ID: cValReg to which value applies

NWCA_CC: state- and taxon-specific Coefficient of Conservatism

If this data frame is not supplied, the data set ccNWCA included in the package is used.

taxaWIS

A data frame containing Wetland Indicator Status as assigned by USAC and reconciled by USDA PLANTS, with the following variables, at a minimum:

USDA_NAME: USDA accepted name

GEOG_ID: USAC region to which value applies

WIS: Wetland Indicator Status

If this data frame is not supplied, the data set wisNWCA included in the package is used.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

state

String containing the name of the state in vascIn, with default value of 'STATE'

coeReg

String containing the name of the U.S. Army Corps of Engineers region in vascIn associated with Wetland Indicator Status, with default value of 'USAC_REGION'

cValReg

String containing the name of the variable in taxaCC which specifies the C-value region.

Value

Either a character string containing an error message when metric calculation is not successful, or a data frame. The first column(s) of the data frame contain the sampID variables, and subsequent columns are named for each metric and contain metric values. A list of vascular plant metrics is installed with the package. Use system.file("Vascular_Plant_Metric_Descriptions.pdf", package="aquametNWCA") to locate the file.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0

exOut <- calcVascPlantMets(VascPlantEx,
  taxon_name = "USDA_NAME",
  taxaIn = taxaNWCA, taxaNat = ccNatNWCA,
  taxaCC = ccNatNWCA, taxaWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "STATE"
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"
#> [1] "Initial datasets prepared for metric calculation."
#> [1] "Done calculating richness metrics."
#> [1] "Ready to start calculating trait metrics."
#> [1] "Done with diversity indices."
#> [1] "Done with duration metrics."
#> [1] "Done with growth habit metrics."
#> [1] "Done with category metrics."
#> [1] "Done with WIS metrics."
#> [1] "Done with CC metrics."
#> [1] "Done with native status metrics."
#> [1] "Done with Bray-Curtis distance."

str(exOut)
#> 'data.frame':	10 obs. of  393 variables:
#>  $ UID                  : int  1 2 3 4 5 6 7 8 9 10
#>  $ XTOTABCOV            : num  203.6 187.6 97.2 110.2 51.5 ...
#>  $ TOTN_SPP             : num  18 49 21 34 6 71 9 4 16 10
#>  $ XN_SPP               : num  9.2 22.4 10.2 15.2 4.8 36 5.8 3.4 8.4 10
#>  $ MEDN_SPP             : num  9 19 10 14 5 40 5 3 8 10
#>  $ SDN_SPP              : num  1.92 8.76 3.77 4.21 0.84 ...
#>  $ TOTN_GEN             : num  18 38 19 30 6 55 9 4 15 10
#>  $ XN_GEN               : num  9.2 20.4 9 14.4 4.8 31.4 5.8 3.4 8.4 10
#>  $ MEDN_GEN             : num  9 18 9 14 5 34 5 3 8 10
#>  $ SDN_GEN              : num  1.92 6.95 3.94 3.78 0.84 ...
#>  $ TOTN_FAM             : num  15 32 11 26 4 31 8 2 9 9
#>  $ XN_FAM               : num  7.4 18.2 6.2 13.4 2.8 22.2 5.8 2 5.2 9
#>  $ MEDN_FAM             : num  8 17 6 13 3 25 5 2 5 9
#>  $ SDN_FAM              : num  1.95 5.97 2.28 3.51 0.84 5.81 1.64 0 0.45 0
#>  $ TOTN_NATSPP          : num  18 45 19 32 6 63 9 3 8 10
#>  $ XN_NATSPP            : num  9.2 21 8.8 14.4 4.8 32.8 5.8 3 4.2 10
#>  $ MEDN_NATSPP          : num  9 19 9 14 5 37 5 3 4 10
#>  $ SDN_NATSPP           : num  1.92 7.18 3.35 4.04 0.84 11.1 1.64 0 0.84 0
#>  $ TOTN_ADVSPP          : num  0 0 0 0 0 0 0 0 1 0
#>  $ XN_ADVSPP            : num  0 0 0 0 0 0 0 0 0.6 0
#>  $ MEDN_ADVSPP          : num  0 0 0 0 0 0 0 0 1 0
#>  $ SDN_ADVSPP           : num  0 0 0 0 0 0 0 0 0.55 0
#>  $ TOTN_CRYPSPP         : num  0 0 0 0 0 1 0 0 0 0
#>  $ XN_CRYPSPP           : num  0 0 0 0 0 0.2 0 0 0 0
#>  $ MEDN_CRYPSPP         : num  0 0 0 0 0 0 0 0 0 0
#>  $ SDN_CRYPSPP          : num  0 0 0 0 0 0.45 0 0 0 0
#>  $ TOTN_INTRSPP         : num  0 4 2 0 0 5 0 1 4 0
#>  $ XN_INTRSPP           : num  0 1.4 1.4 0 0 2 0 0.4 1.8 0
#>  $ MEDN_INTRSPP         : num  0 1 1 0 0 2 0 0 2 0
#>  $ SDN_INTRSPP          : num  0 1.67 0.55 0 0 1.58 0 0.55 0.84 0
#>  $ TOTN_ALIENSPP        : num  0 4 2 0 0 5 0 1 5 0
#>  $ XN_ALIENSPP          : num  0 1.4 1.4 0 0 2 0 0.4 2.4 0
#>  $ MEDN_ALIENSPP        : num  0 1 1 0 0 2 0 0 3 0
#>  $ SDN_ALIENSPP         : num  0 1.67 0.55 0 0 1.58 0 0.55 1.34 0
#>  $ TOTN_AC              : num  0 4 2 0 0 6 0 1 5 0
#>  $ XN_AC                : num  0 1.4 1.4 0 0 2.2 0 0.4 2.4 0
#>  $ MEDN_AC              : num  0 1 1 0 0 2 0 0 3 0
#>  $ SDN_AC               : num  0 1.67 0.55 0 0 1.79 0 0.55 1.34 0
#>  $ H_ALL                : num  1.948 2.057 1.7 3.124 0.615 ...
#>  $ J_ALL                : num  0.674 0.528 0.558 0.886 0.343 ...
#>  $ D_ALL                : num  0.82 0.745 0.78 0.946 0.292 ...
#>  $ H_NAT                : num  1.948 2.003 1.522 3.09 0.615 ...
#>  $ J_NAT                : num  0.674 0.526 0.517 0.892 0.343 ...
#>  $ D_NAT                : num  0.82 0.739 0.739 0.944 0.292 ...
#>  $ H_ALIEN              : num  0 1.121 0.026 0 0 ...
#>  $ J_ALIEN              : num  0 0.8086 0.0375 0 0 ...
#>  $ D_ALIEN              : num  0 0.6116 0.0079 0 0 ...
#>  $ H_AC                 : num  0 1.121 0.026 0 0 ...
#>  $ J_AC                 : num  0 0.8086 0.0375 0 0 ...
#>  $ D_AC                 : num  0 0.6116 0.0079 0 0 ...
#>  $ N_PERENNIAL          : num  18 46 20 30 5 64 8 4 7 9
#>  $ N_ANNUAL             : num  0 2 0 1 0 2 0 0 4 0
#>  $ N_ANN_PEREN          : num  0 1 1 1 1 3 0 0 2 0
#>  $ N_ANN_BIEN           : num  0 0 0 0 0 0 1 0 0 0
#>  $ PCTN_PERENNIAL       : num  100 93.9 95.2 88.2 83.3 ...
#>  $ PCTN_ANNUAL          : num  0 4.08 0 2.94 0 2.82 0 0 25 0
#>  $ PCTN_ANN_PEREN       : num  0 2.04 4.76 2.94 16.67 ...
#>  $ PCTN_ANN_BIEN        : num  0 0 0 0 0 ...
#>  $ XABCOV_PERENNIAL     : num  203.6 185.6 97.2 107.2 50.7 ...
#>  $ XABCOV_ANNUAL        : num  0 1.6 0 0.6 0 1.66 0 0 24.4 0
#>  $ XABCOV_ANN_PEREN     : num  0 0.4 0.04 1 0.8 ...
#>  $ XABCOV_ANN_BIEN      : num  0 0 0 0 0 0 0.6 0 0 0
#>  $ XRCOV_PERENNIAL      : num  100 98.9 100 97.3 98.5 ...
#>  $ XRCOV_ANNUAL         : num  0 0.85 0 0.54 0 ...
#>  $ XRCOV_ANN_PEREN      : num  0 0.21 0.04 0.91 1.55 ...
#>  $ XRCOV_ANN_BIEN       : num  0 0 0 0 0 0 0.58 0 0 0
#>  $ N_ANNUAL_NAT         : num  0 1 0 1 0 2 0 0 2 0
#>  $ PCTN_ANNUAL_NAT      : num  0 2.04 0 2.94 0 2.82 0 0 12.5 0
#>  $ XABCOV_ANNUAL_NAT    : num  0 0.4 0 0.6 0 ...
#>  $ XRCOV_ANNUAL_NAT     : num  0 0.21 0 0.54 0 ...
#>  $ N_ANNUAL_AC          : num  0 1 0 0 0 0 0 0 2 0
#>  $ PCTN_ANNUAL_AC       : num  0 2.04 0 0 0 0 0 0 12.5 0
#>  $ XABCOV_ANNUAL_AC     : num  0 1.2 0 0 0 0 0 0 9.32 0
#>  $ XRCOV_ANNUAL_AC      : num  0 0.64 0 0 0 0 0 0 7.66 0
#>  $ N_ANN_BIEN_NAT       : num  0 0 0 0 0 0 1 0 0 0
#>  $ PCTN_ANN_BIEN_NAT    : num  0 0 0 0 0 ...
#>  $ XABCOV_ANN_BIEN_NAT  : num  0 0 0 0 0 0 0.6 0 0 0
#>  $ XRCOV_ANN_BIEN_NAT   : num  0 0 0 0 0 0 0.58 0 0 0
#>  $ N_ANN_BIEN_AC        : num  0 0 0 0 0 0 0 0 0 0
#>  $ PCTN_ANN_BIEN_AC     : num  0 0 0 0 0 0 0 0 0 0
#>  $ XABCOV_ANN_BIEN_AC   : num  0 0 0 0 0 0 0 0 0 0
#>  $ XRCOV_ANN_BIEN_AC    : num  0 0 0 0 0 0 0 0 0 0
#>  $ N_ANN_PEREN_NAT      : num  0 1 1 1 1 3 0 0 0 0
#>  $ PCTN_ANN_PEREN_NAT   : num  0 2.04 4.76 2.94 16.67 ...
#>  $ XABCOV_ANN_PEREN_NAT : num  0 0.4 0.04 1 0.8 0.14 0 0 0 0
#>  $ XRCOV_ANN_PEREN_NAT  : num  0 0.21 0.04 0.91 1.55 0.12 0 0 0 0
#>  $ N_ANN_PEREN_AC       : num  0 0 0 0 0 0 0 0 2 0
#>  $ PCTN_ANN_PEREN_AC    : num  0 0 0 0 0 0 0 0 12.5 0
#>  $ XABCOV_ANN_PEREN_AC  : num  0 0 0 0 0 ...
#>  $ XRCOV_ANN_PEREN_AC   : num  0 0 0 0 0 ...
#>  $ N_PERENNIAL_NAT      : num  18 43 18 30 5 58 8 3 6 9
#>  $ PCTN_PERENNIAL_NAT   : num  100 87.8 85.7 88.2 83.3 ...
#>  $ XABCOV_PERENNIAL_NAT : num  203.6 184.6 87.1 107.2 50.7 ...
#>  $ XRCOV_PERENNIAL_NAT  : num  100 98.4 89.6 97.3 98.5 ...
#>  $ N_PERENNIAL_AC       : num  0 3 2 0 0 6 0 1 1 0
#>  $ PCTN_PERENNIAL_AC    : num  0 6.12 9.52 0 0 8.45 0 25 6.25 0
#>  $ XABCOV_PERENNIAL_AC  : num  0 1 10 0 0 ...
#>  $ XRCOV_PERENNIAL_AC   : num  0 0.53 10.33 0 0 ...
#>  $ N_FORB               : num  1 18 10 9 1 30 4 0 4 6
#>   [list output truncated]
#>  - attr(*, "reshapeWide")=List of 5
#>   ..$ v.names: chr "RESULT"
#>   ..$ timevar: chr "PARAMETER"
#>   ..$ idvar  : chr "UID"
#>   ..$ times  : chr [1:392] "XTOTABCOV" "TOTN_SPP" "XN_SPP" "MEDN_SPP" ...
#>   ..$ varying: chr [1, 1:392] "RESULT.XTOTABCOV" "RESULT.TOTN_SPP" "RESULT.XN_SPP" "RESULT.MEDN_SPP" ...
head(exOut)
#>             UID XTOTABCOV TOTN_SPP XN_SPP MEDN_SPP SDN_SPP TOTN_GEN XN_GEN
#> 1.XTOTABCOV   1    203.56       18    9.2        9    1.92       18    9.2
#> 2.XTOTABCOV   2    187.60       49   22.4       19    8.76       38   20.4
#>             MEDN_GEN SDN_GEN TOTN_FAM XN_FAM MEDN_FAM SDN_FAM TOTN_NATSPP
#> 1.XTOTABCOV        9    1.92       15    7.4        8    1.95          18
#> 2.XTOTABCOV       18    6.95       32   18.2       17    5.97          45
#>             XN_NATSPP MEDN_NATSPP SDN_NATSPP TOTN_ADVSPP XN_ADVSPP MEDN_ADVSPP
#> 1.XTOTABCOV       9.2           9       1.92           0         0           0
#> 2.XTOTABCOV      21.0          19       7.18           0         0           0
#>             SDN_ADVSPP TOTN_CRYPSPP XN_CRYPSPP MEDN_CRYPSPP SDN_CRYPSPP
#> 1.XTOTABCOV          0            0          0            0           0
#> 2.XTOTABCOV          0            0          0            0           0
#>             TOTN_INTRSPP XN_INTRSPP MEDN_INTRSPP SDN_INTRSPP TOTN_ALIENSPP
#> 1.XTOTABCOV            0        0.0            0        0.00             0
#> 2.XTOTABCOV            4        1.4            1        1.67             4
#>             XN_ALIENSPP MEDN_ALIENSPP SDN_ALIENSPP TOTN_AC XN_AC MEDN_AC SDN_AC
#> 1.XTOTABCOV         0.0             0         0.00       0   0.0       0   0.00
#> 2.XTOTABCOV         1.4             1         1.67       4   1.4       1   1.67
#>              H_ALL  J_ALL  D_ALL  H_NAT  J_NAT  D_NAT H_ALIEN J_ALIEN D_ALIEN
#> 1.XTOTABCOV 1.9476 0.6738 0.8196 1.9476 0.6738 0.8196   0.000  0.0000  0.0000
#> 2.XTOTABCOV 2.0566 0.5284 0.7449 2.0031 0.5262 0.7388   1.121  0.8086  0.6116
#>              H_AC   J_AC   D_AC N_PERENNIAL N_ANNUAL N_ANN_PEREN N_ANN_BIEN
#> 1.XTOTABCOV 0.000 0.0000 0.0000          18        0           0          0
#> 2.XTOTABCOV 1.121 0.8086 0.6116          46        2           1          0
#>             PCTN_PERENNIAL PCTN_ANNUAL PCTN_ANN_PEREN PCTN_ANN_BIEN
#> 1.XTOTABCOV         100.00        0.00           0.00             0
#> 2.XTOTABCOV          93.88        4.08           2.04             0
#>             XABCOV_PERENNIAL XABCOV_ANNUAL XABCOV_ANN_PEREN XABCOV_ANN_BIEN
#> 1.XTOTABCOV           203.56           0.0              0.0               0
#> 2.XTOTABCOV           185.60           1.6              0.4               0
#>             XRCOV_PERENNIAL XRCOV_ANNUAL XRCOV_ANN_PEREN XRCOV_ANN_BIEN
#> 1.XTOTABCOV          100.00         0.00            0.00              0
#> 2.XTOTABCOV           98.93         0.85            0.21              0
#>             N_ANNUAL_NAT PCTN_ANNUAL_NAT XABCOV_ANNUAL_NAT XRCOV_ANNUAL_NAT
#> 1.XTOTABCOV            0            0.00               0.0             0.00
#> 2.XTOTABCOV            1            2.04               0.4             0.21
#>             N_ANNUAL_AC PCTN_ANNUAL_AC XABCOV_ANNUAL_AC XRCOV_ANNUAL_AC
#> 1.XTOTABCOV           0           0.00              0.0            0.00
#> 2.XTOTABCOV           1           2.04              1.2            0.64
#>             N_ANN_BIEN_NAT PCTN_ANN_BIEN_NAT XABCOV_ANN_BIEN_NAT
#> 1.XTOTABCOV              0                 0                   0
#> 2.XTOTABCOV              0                 0                   0
#>             XRCOV_ANN_BIEN_NAT N_ANN_BIEN_AC PCTN_ANN_BIEN_AC
#> 1.XTOTABCOV                  0             0                0
#> 2.XTOTABCOV                  0             0                0
#>             XABCOV_ANN_BIEN_AC XRCOV_ANN_BIEN_AC N_ANN_PEREN_NAT
#> 1.XTOTABCOV                  0                 0               0
#> 2.XTOTABCOV                  0                 0               1
#>             PCTN_ANN_PEREN_NAT XABCOV_ANN_PEREN_NAT XRCOV_ANN_PEREN_NAT
#> 1.XTOTABCOV               0.00                  0.0                0.00
#> 2.XTOTABCOV               2.04                  0.4                0.21
#>             N_ANN_PEREN_AC PCTN_ANN_PEREN_AC XABCOV_ANN_PEREN_AC
#> 1.XTOTABCOV              0                 0                   0
#> 2.XTOTABCOV              0                 0                   0
#>             XRCOV_ANN_PEREN_AC N_PERENNIAL_NAT PCTN_PERENNIAL_NAT
#> 1.XTOTABCOV                  0              18             100.00
#> 2.XTOTABCOV                  0              43              87.76
#>             XABCOV_PERENNIAL_NAT XRCOV_PERENNIAL_NAT N_PERENNIAL_AC
#> 1.XTOTABCOV               203.56               100.0              0
#> 2.XTOTABCOV               184.60                98.4              3
#>             PCTN_PERENNIAL_AC XABCOV_PERENNIAL_AC XRCOV_PERENNIAL_AC N_FORB
#> 1.XTOTABCOV              0.00                   0               0.00      1
#> 2.XTOTABCOV              6.12                   1               0.53     18
#>             N_GRAMINOID N_SHRUB N_TREE N_TREE_SHRUB N_VINE N_VINE_SHRUB
#> 1.XTOTABCOV           1       4      6            4      1            1
#> 2.XTOTABCOV          10       2      6            5      5            2
#>             N_SSHRUB_FORB N_SSHRUB_SHRUB PCTN_FORB PCTN_GRAMINOID PCTN_SHRUB
#> 1.XTOTABCOV             0              0      5.56           5.56      22.22
#> 2.XTOTABCOV             1              0     36.73          20.41       4.08
#>             PCTN_TREE PCTN_TREE_SHRUB PCTN_VINE PCTN_VINE_SHRUB
#> 1.XTOTABCOV     33.33           22.22      5.56            5.56
#> 2.XTOTABCOV     12.24           10.20     10.20            4.08
#>             PCTN_SSHRUB_FORB PCTN_SSHRUB_SHRUB XABCOV_FORB XABCOV_GRAMINOID
#> 1.XTOTABCOV             0.00                 0         0.4             0.02
#> 2.XTOTABCOV             2.04                 0        91.4             5.40
#>             XABCOV_SHRUB XABCOV_TREE XABCOV_TREE_SHRUB XABCOV_VINE
#> 1.XTOTABCOV        57.62      137.66              6.64        0.02
#> 2.XTOTABCOV         1.40       55.00             28.60        2.20
#>             XABCOV_VINE_SHRUB XABCOV_SSHRUB_FORB XABCOV_SSHRUB_SHRUB XRCOV_FORB
#> 1.XTOTABCOV               1.2                0.0                   0       0.20
#> 2.XTOTABCOV               3.4                0.2                   0      48.72
#>             XRCOV_GRAMINOID XRCOV_SHRUB XRCOV_TREE XRCOV_TREE_SHRUB XRCOV_VINE
#> 1.XTOTABCOV            0.01       28.31      67.63             3.26       0.01
#> 2.XTOTABCOV            2.88        0.75      29.32            15.25       1.17
#>             XRCOV_VINE_SHRUB XRCOV_SSHRUB_FORB XRCOV_SSHRUB_SHRUB N_TREE_COMB
#> 1.XTOTABCOV             0.59              0.00                  0          10
#> 2.XTOTABCOV             1.81              0.11                  0          11
#>             PCTN_TREE_COMB XABCOV_TREE_COMB XRCOV_TREE_COMB N_SHRUB_COMB
#> 1.XTOTABCOV          55.56            144.3           70.89            4
#> 2.XTOTABCOV          22.45             83.6           44.56            2
#>             PCTN_SHRUB_COMB XABCOV_SHRUB_COMB XRCOV_SHRUB_COMB N_HERB PCTN_HERB
#> 1.XTOTABCOV           22.22             57.62            28.31      2     11.11
#> 2.XTOTABCOV            4.08              1.40             0.75     29     59.18
#>             XABCOV_HERB XRCOV_HERB N_VINE_ALL PCTN_VINE_ALL XABCOV_VINE_ALL
#> 1.XTOTABCOV        0.42       0.21          2         11.11            1.22
#> 2.XTOTABCOV       97.00      51.71          7         14.29            5.60
#>             XRCOV_VINE_ALL N_GRAMINOID_AC PCTN_GRAMINOID_AC XABCOV_GRAMINOID_AC
#> 1.XTOTABCOV           0.60              0              0.00                 0.0
#> 2.XTOTABCOV           2.99              1              2.04                 1.2
#>             XRCOV_GRAMINOID_AC N_GRAMINOID_NAT XRCOV_GRAMINOID_NAT
#> 1.XTOTABCOV               0.00               1                0.01
#> 2.XTOTABCOV               0.64               9                2.24
#>             XABCOV_GRAMINOID_NAT PCTN_GRAMINOID_NAT N_FORB_AC PCTN_FORB_AC
#> 1.XTOTABCOV                 0.02               5.56         0         0.00
#> 2.XTOTABCOV                 4.20              18.37         1         2.04
#>             XABCOV_FORB_AC XRCOV_FORB_AC N_FORB_NAT XRCOV_FORB_NAT
#> 1.XTOTABCOV            0.0          0.00          1           0.20
#> 2.XTOTABCOV            0.2          0.11         17          48.61
#>             XABCOV_FORB_NAT PCTN_FORB_NAT N_HERB_AC PCTN_HERB_AC XABCOV_HERB_AC
#> 1.XTOTABCOV             0.4          5.56         0         0.00            0.0
#> 2.XTOTABCOV            91.2         34.69         3         6.12            1.6
#>             XRCOV_HERB_AC N_HERB_NAT PCTN_HERB_NAT XABCOV_HERB_NAT
#> 1.XTOTABCOV          0.00          2         11.11            0.42
#> 2.XTOTABCOV          0.85         26         53.06           95.40
#>             XRCOV_HERB_NAT N_SHRUB_COMB_AC PCTN_SHRUB_COMB_AC
#> 1.XTOTABCOV           0.21               0                  0
#> 2.XTOTABCOV          50.85               0                  0
#>             XABCOV_SHRUB_COMB_AC XRCOV_SHRUB_COMB_AC N_SHRUB_COMB_NAT
#> 1.XTOTABCOV                    0                   0                4
#> 2.XTOTABCOV                    0                   0                2
#>             PCTN_SHRUB_COMB_NAT XABCOV_SHRUB_COMB_NAT XRCOV_SHRUB_COMB_NAT
#> 1.XTOTABCOV               22.22                 57.62                28.31
#> 2.XTOTABCOV                4.08                  1.40                 0.75
#>             N_TREE_COMB_AC PCTN_TREE_COMB_AC XABCOV_TREE_COMB_AC
#> 1.XTOTABCOV              0                 0                   0
#> 2.XTOTABCOV              0                 0                   0
#>             XRCOV_TREE_COMB_AC N_TREE_COMB_NAT PCTN_TREE_COMB_NAT
#> 1.XTOTABCOV                  0              10              55.56
#> 2.XTOTABCOV                  0              11              22.45
#>             XABCOV_TREE_COMB_NAT XRCOV_TREE_COMB_NAT N_VINE_AC PCTN_VINE_AC
#> 1.XTOTABCOV                144.3               70.89         0         0.00
#> 2.XTOTABCOV                 83.6               44.56         1         2.04
#>             XABCOV_VINE_AC XRCOV_VINE_AC N_VINE_NAT PCTN_VINE_NAT
#> 1.XTOTABCOV            0.0          0.00          1          5.56
#> 2.XTOTABCOV            0.6          0.32          4          8.16
#>             XABCOV_VINE_NAT XRCOV_VINE_NAT N_VINE_SHRUB_AC PCTN_VINE_SHRUB_AC
#> 1.XTOTABCOV            0.02           0.01               0                  0
#> 2.XTOTABCOV            1.60           0.85               0                  0
#>             XABCOV_VINE_SHRUB_AC XRCOV_VINE_SHRUB_AC N_VINE_SHRUB_NAT
#> 1.XTOTABCOV                    0                   0                1
#> 2.XTOTABCOV                    0                   0                2
#>             PCTN_VINE_SHRUB_NAT XABCOV_VINE_SHRUB_NAT XRCOV_VINE_SHRUB_NAT
#> 1.XTOTABCOV                5.56                   1.2                 0.59
#> 2.XTOTABCOV                4.08                   3.4                 1.81
#>             N_VINE_ALL_AC PCTN_VINE_ALL_AC XABCOV_VINE_ALL_AC XRCOV_VINE_ALL_AC
#> 1.XTOTABCOV             0             0.00                0.0              0.00
#> 2.XTOTABCOV             1             2.04                0.6              0.32
#>             N_VINE_ALL_NAT PCTN_VINE_ALL_NAT XABCOV_VINE_ALL_NAT
#> 1.XTOTABCOV              2             11.11                1.22
#> 2.XTOTABCOV              6             12.24                5.00
#>             XRCOV_VINE_ALL_NAT N_DICOT N_FERN N_MONOCOT N_GYMNOSPERM
#> 1.XTOTABCOV               0.60      15      1         2            0
#> 2.XTOTABCOV               2.67      29      5        15            0
#>             N_HORSETAIL PCTN_DICOT PCTN_FERN PCTN_MONOCOT PCTN_GYMNOSPERM
#> 1.XTOTABCOV           0      83.33      5.56        11.11               0
#> 2.XTOTABCOV           0      59.18     10.20        30.61               0
#>             PCTN_HORSETAIL XABCOV_DICOT XABCOV_FERN XABCOV_MONOCOT
#> 1.XTOTABCOV              0       201.94         0.4           1.22
#> 2.XTOTABCOV              0        92.20        87.2           8.20
#>             XABCOV_GYMNOSPERM XABCOV_HORSETAIL XRCOV_DICOT XRCOV_FERN
#> 1.XTOTABCOV                 0                0       99.20       0.20
#> 2.XTOTABCOV                 0                0       49.15      46.48
#>             XRCOV_MONOCOT XRCOV_GYMNOSPERM XRCOV_HORSETAIL N_DICOTS_NAT
#> 1.XTOTABCOV          0.60                0               0           15
#> 2.XTOTABCOV          4.37                0               0           26
#>             XRCOV_DICOTS_NAT XABCOV_DICOTS_NAT PCTN_DICOTS_NAT N_DICOTS_ALIEN
#> 1.XTOTABCOV            99.20            201.94           83.33              0
#> 2.XTOTABCOV            48.61             91.20           53.06              3
#>             PCTN_DICOTS_ALIEN XABCOV_DICOTS_ALIEN XRCOV_DICOTS_ALIEN
#> 1.XTOTABCOV              0.00                   0               0.00
#> 2.XTOTABCOV              6.12                   1               0.53
#>             N_DICOTS_CRYP PCTN_DICOTS_CRYP XABCOV_DICOTS_CRYP XRCOV_DICOTS_CRYP
#> 1.XTOTABCOV             0                0                  0                 0
#> 2.XTOTABCOV             0                0                  0                 0
#>             N_DICOTS_AC PCTN_DICOTS_AC XABCOV_DICOTS_AC XRCOV_DICOTS_AC
#> 1.XTOTABCOV           0           0.00                0            0.00
#> 2.XTOTABCOV           3           6.12                1            0.53
#>             N_FERNS_NAT XRCOV_FERNS_NAT XABCOV_FERNS_NAT PCTN_FERNS_NAT
#> 1.XTOTABCOV           1            0.20              0.4           5.56
#> 2.XTOTABCOV           5           46.48             87.2          10.20
#>             N_FERNS_INTR PCTN_FERNS_INTR XABCOV_FERNS_INTR XRCOV_FERNS_INTR
#> 1.XTOTABCOV            0               0                 0                0
#> 2.XTOTABCOV            0               0                 0                0
#>             N_MONOCOTS_NAT XRCOV_MONOCOTS_NAT XABCOV_MONOCOTS_NAT
#> 1.XTOTABCOV              2               0.60                1.22
#> 2.XTOTABCOV             14               3.73                7.00
#>             PCTN_MONOCOTS_NAT N_MONOCOTS_ALIEN PCTN_MONOCOTS_ALIEN
#> 1.XTOTABCOV             11.11                0                0.00
#> 2.XTOTABCOV             28.57                1                2.04
#>             XABCOV_MONOCOTS_ALIEN XRCOV_MONOCOTS_ALIEN N_MONOCOTS_CRYP
#> 1.XTOTABCOV                   0.0                 0.00               0
#> 2.XTOTABCOV                   1.2                 0.64               0
#>             PCTN_MONOCOTS_CRYP XABCOV_MONOCOTS_CRYP XRCOV_MONOCOTS_CRYP
#> 1.XTOTABCOV                  0                    0                   0
#> 2.XTOTABCOV                  0                    0                   0
#>             N_MONOCOTS_AC PCTN_MONOCOTS_AC XABCOV_MONOCOTS_AC XRCOV_MONOCOTS_AC
#> 1.XTOTABCOV             0             0.00                0.0              0.00
#> 2.XTOTABCOV             1             2.04                1.2              0.64
#>             N_LYCOPOD N_QUILLWORT N_LIVERWORT PCTN_LYCOPOD PCTN_QUILLWORT
#> 1.XTOTABCOV         0           0           0            0              0
#> 2.XTOTABCOV         0           0           0            0              0
#>             PCTN_LIVERWORT XABCOV_LYCOPOD XABCOV_QUILLWORT XABCOV_LIVERWORT
#> 1.XTOTABCOV              0              0                0                0
#> 2.XTOTABCOV              0              0                0                0
#>             XRCOV_LYCOPOD XRCOV_QUILLWORT XRCOV_LIVERWORT N_FAC N_FACU N_FACW
#> 1.XTOTABCOV             0               0               0     8      1      7
#> 2.XTOTABCOV             0               0               0    12      6     20
#>             N_OBL N_UPL PCTN_FAC PCTN_FACU PCTN_FACW PCTN_OBL PCTN_UPL
#> 1.XTOTABCOV     1     0    44.44      5.56     38.89     5.56     0.00
#> 2.XTOTABCOV     8     2    24.49     12.24     40.82    16.33     4.08
#>             XABCOV_FAC XABCOV_FACU XABCOV_FACW XABCOV_OBL XABCOV_UPL XRCOV_FAC
#> 1.XTOTABCOV     143.64        0.02       56.84       3.04        0.0     70.56
#> 2.XTOTABCOV     133.60       38.40       11.80       3.00        0.6     71.22
#>             XRCOV_FACU XRCOV_FACW XRCOV_OBL XRCOV_UPL N_OBL_FACW PCTN_OBL_FACW
#> 1.XTOTABCOV       0.01      27.92      1.49      0.00          8         44.44
#> 2.XTOTABCOV      20.47       6.29      1.60      0.32         28         57.14
#>             XABCOV_OBL_FACW XRCOV_OBL_FACW N_OBL_FACW_FAC PCTN_OBL_FACW_FAC
#> 1.XTOTABCOV           59.88          29.42             16             88.89
#> 2.XTOTABCOV           14.80           7.89             40             81.63
#>             XABCOV_OBL_FACW_FAC XRCOV_OBL_FACW_FAC N_FAC_FACU PCTN_FAC_FACU
#> 1.XTOTABCOV              203.52              99.98          9         50.00
#> 2.XTOTABCOV              148.40              79.10         18         36.73
#>             XABCOV_FAC_FACU XRCOV_FAC_FACU WETIND_COV_ALL WETIND_FREQ_ALL
#> 1.XTOTABCOV          143.66          70.57           2.69            2.49
#> 2.XTOTABCOV          172.00          91.68           3.12            2.52
#>             WETIND2_COV_ALL WETIND2_FREQ_ALL WETIND_COV_NAT WETIND_FREQ_NAT
#> 1.XTOTABCOV            3.31             3.51           2.69            2.49
#> 2.XTOTABCOV            2.88             3.48           3.11            2.46
#>             WETIND2_COV_NAT WETIND2_FREQ_NAT N_OBLFACW_AC XRCOV_OBLFACW_AC
#> 1.XTOTABCOV            3.31             3.51            0                0
#> 2.XTOTABCOV            2.89             3.54            0                0
#>             XABCOV_OBLFACW_AC XC_ALL XC_FREQ_ALL XC_COV_ALL FQAI_ALL
#> 1.XTOTABCOV                 0   4.78       23.67      18.84    20.27
#> 2.XTOTABCOV                 0   4.49        8.80       9.34    31.43
#>             FQAI_FREQ_ALL FQAI_COV_ALL N_SEN XRCOV_SEN XABCOV_SEN PCTN_SEN
#> 1.XTOTABCOV        100.43        79.94     1      0.01       0.02     5.56
#> 2.XTOTABCOV         61.61        65.38     5      0.85       1.60    10.20
#>             N_TOL XRCOV_TOL XABCOV_TOL PCTN_TOL N_ISEN XRCOV_ISEN XABCOV_ISEN
#> 1.XTOTABCOV     8     78.24     159.26    44.44      9      21.75       44.28
#> 2.XTOTABCOV    16     21.22      39.80    32.65     28      77.93      146.20
#>             PCTN_ISEN N_HTOL XRCOV_HTOL XABCOV_HTOL PCTN_HTOL N_HSEN XRCOV_HSEN
#> 1.XTOTABCOV     50.00      3      52.18      106.22     16.67      0          0
#> 2.XTOTABCOV     57.14      9      12.37       23.20     18.37      0          0
#>             XABCOV_HSEN PCTN_HSEN XC_NAT XC_FREQ_NAT XC_COV_NAT FQAI_NAT
#> 1.XTOTABCOV           0         0   4.78       23.67      18.84    20.27
#> 2.XTOTABCOV           0         0   4.89       10.22      10.29    32.80
#>             FQAI_FREQ_NAT FQAI_COV_NAT PCTN_NATSPP PCTN_INTRSPP PCTN_CRYPSPP
#> 1.XTOTABCOV        100.43        79.94      100.00         0.00            0
#> 2.XTOTABCOV         68.57        69.04       91.84         8.16            0
#>             PCTN_ADVSPP XABCOV_NATSPP XABCOV_INTRSPP XABCOV_CRYPSPP
#> 1.XTOTABCOV           0        203.56            0.0              0
#> 2.XTOTABCOV           0        185.40            2.2              0
#>             XABCOV_ADVSPP XRCOV_NATSPP XRCOV_INTRSPP XRCOV_CRYPSPP XRCOV_ADVSPP
#> 1.XTOTABCOV             0       100.00          0.00             0            0
#> 2.XTOTABCOV             0        98.83          1.17             0            0
#>             RFREQ_NATSPP RFREQ_INTRSPP RFREQ_CRYPSPP RFREQ_ADVSPP RIMP_NATSPP
#> 1.XTOTABCOV       100.00          0.00             0            0      100.00
#> 2.XTOTABCOV        93.75          6.25             0            0       96.29
#>             RIMP_INTRSPP RIMP_CRYPSPP RIMP_ADVSPP PCTN_ALIENSPP XABCOV_ALIENSPP
#> 1.XTOTABCOV         0.00            0           0          0.00             0.0
#> 2.XTOTABCOV         3.71            0           0          8.16             2.2
#>             XRCOV_ALIENSPP RFREQ_ALIENSPP RIMP_ALIENSPP PCTN_AC XABCOV_AC
#> 1.XTOTABCOV           0.00           0.00          0.00    0.00       0.0
#> 2.XTOTABCOV           1.17           6.25          3.71    8.16       2.2
#>             XRCOV_AC RFREQ_AC RIMP_AC XBCDIST_SPP XBCDIST_NATSPP
#> 1.XTOTABCOV     0.00     0.00    0.00      0.5080         0.5080
#> 2.XTOTABCOV     1.17     6.25    3.71      0.3721         0.3668
#>  [ reached 'max' / getOption("max.print") -- omitted 4 rows ]