Calculate NWCA 2011 vegetation types and ground surface attribute metrics based on data collected from field form V-3

This function calculates the NWCA 2011 vegetation types and ground surface attribute metrics. It assumes input data are organized so that each row represents the cover for a single taxon within a single plot at a site. It also assumes a single column is used to identify each unique sample. Missing values and blanks should be removed from the input data frame.

calcVtype_GcovMets(dataIn, nPlotIn, sampID = "UID", survyear = "2011")

Arguments

dataIn

A data frame containing the following variables:

sampID: Variables identified in the sampID argument

PLOT: Sample plot from which data were collected

PARAMETER: specific measurement type

RESULT: measured value

The following parameters are used in calculating tree metrics: 'SANDT_CLASS','SUBMERGED_AQ', 'FLOATING_AQ','LIANAS','VTALL_VEG','TALL_VEG','HMED_VEG', 'MED_VEG','SMALL_VEG','VSMALL_VEG','PEAT_MOSS','BRYOPHYTES', 'LICHENS','ARBOREAL','ALGAE','MACROALGAE','TIME','MINIMUM_DEPTH', 'MAXIMUM_DEPTH','PREDOMINANT_DEPTH','TOTAL_WATER','WATER_NOVEG', 'WATER_AQVEG','WATER_EMERGVEG','LITTER_THATCH','LITTER_FORB', 'LITTER_CONIFER','LITTER_DECID','LITTER_BROADLEAF', 'LITTER_DEPTH_SW','LITTER_DEPTH_NE','TOTAL_LITTER','WD_FINE', 'WD_COARSE','EXPOSED_SOIL','EXPOSED_GRAVEL','EXPOSED_ROCK', 'EXPOSED_SOIL','EXPOSED_GRAVEL','EXPOSED_ROCK','WD_FINE', 'WD_COARSE','TOTAL_LITTER'. Additional parameters or variables are ignored.

nPlotIn

A data frame with the number of plots sampled associated with each sample, with sampID variables and NPLOTS.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

survyear

A string with the survey year. Some parameters only measured in 2011, applies to water cover metrics and bare ground and litter metrics.

Value

Either a character string containing an error message when metric calculation is not successful, or a data frame. The first column of the data frame is the sampID variables and subsequent columns are named for each metric and contain metric values. A list of vegetation type and ground cover metrics is installed with the package. Use system.file("VegTypes_GrdCover_Metric_Descriptions.pdf", package="aquametNWCA") to locate the file.

Details

If any of the parameters are missing, they are assumed to be zeros (if numeric), and metric values associated with any metrics that cannot be calculated due to missing parameters are set to a standardized value.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(Vtype_GrCovEx)
#>   UID PLOT       PARAMETER RESULT
#> 1   7    4   MINIMUM_DEPTH      8
#> 2   7    3 LITTER_DEPTH_NE      0
#> 3   7    5        TALL_VEG      0
#> 4   7    4     WATER_AQVEG      0
#> 5   7    5   MINIMUM_DEPTH      8
#> 6   7    5     WATER_NOVEG      0

subVtype_GrCovEx <- subset(Vtype_GrCovEx, PARAMETER != "SANDT_CLASS")

nplots <- aggregate(
  x = list(NPLOTS = subVtype_GrCovEx$PLOT),
  by = subVtype_GrCovEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

exOut <- calcVtype_GcovMets(Vtype_GrCovEx, nPlotIn = nplots, sampID = "UID")
#> [1] "Done with veg heterogeneity metrics"
#> [1] "Done with vascular strata metrics"
#> [1] "Done with non-vascular strata metrics"
#> [1] "Done with water depth metrics"
#> [1] "Done with litter metrics"
#> [1] "Done with bare ground metrics"

str(exOut)
#> 'data.frame':	10 obs. of  109 variables:
#>  $ UID                : int  1 2 3 4 5 6 7 8 9 10
#>  $ D_SANDT            : chr  "0" "0" "0" "0" ...
#>  $ N_SANDT            : chr  "1" "1" "1" "1" ...
#>  $ H_SANDT            : chr  "0" "0" "0" "0" ...
#>  $ J_SANDT            : chr  "0" "0" "0" "0" ...
#>  $ DOM_SANDT          : chr  "PFO" "PFO" "PEM" "PFO" ...
#>  $ FREQ_HMED_VEG      : chr  "80" "100" "0" "100" ...
#>  $ N_VASC_STRATA      : chr  "6" "6" "3" "6" ...
#>  $ IMP_SMALL_VEG      : chr  "70.5" "41" "80.8" "52.2" ...
#>  $ XCOV_FLOATING_AQ   : chr  "0" "0" "0" "0" ...
#>  $ XRCOV_VTALL_VEG    : chr  "0" "0" "0" "0" ...
#>  $ IMP_MED_VEG        : chr  "62.5" "52.8" "62.4" "65.5" ...
#>  $ XCOV_SUBMERGED_AQ  : chr  "0" "0" "0" "0" ...
#>  $ XCOV_VTALL_VEG     : chr  "0" "0" "0" "0" ...
#>  $ IMP_LIANAS         : chr  "30.5" "51.2" "0" "53.2" ...
#>  $ XCOV_VSMALL_VEG    : chr  "4.22" "85" "0.6" "23" ...
#>  $ FREQ_LIANAS        : chr  "60" "100" "0" "100" ...
#>  $ XTOTCOV_VASC_STRATA: chr  "176.22" "140.4" "87" "152.8" ...
#>  $ IMP_TALL_VEG       : chr  "94.5" "66.5" "0" "83.5" ...
#>  $ XRCOV_HMED_VEG     : chr  "9.08" "8.83" "0" "13.74" ...
#>  $ FREQ_TALL_VEG      : chr  "100" "100" "0" "100" ...
#>  $ XCOV_HMED_VEG      : chr  "16" "12.4" "0" "21" ...
#>  $ IMP_FLOATING_AQ    : chr  "0" "0" "0" "0" ...
#>  $ FREQ_SMALL_VEG     : chr  "100" "80" "100" "100" ...
#>  $ RG_VASC_STRATA     : chr  "2" "1" "1" "1" ...
#>  $ IMP_SUBMERGED_AQ   : chr  "0" "0" "0" "0" ...
#>  $ FREQ_MED_VEG       : chr  "100" "100" "100" "100" ...
#>  $ XN_VASC_STRATA     : chr  "5.2" "5.8" "2.6" "5.8" ...
#>  $ IMP_VSMALL_VEG     : chr  "42.11" "92.5" "30.3" "51.5" ...
#>  $ XRCOV_LIANAS       : chr  "0.57" "1.71" "0" "4.19" ...
#>  $ FREQ_VSMALL_VEG    : chr  "80" "100" "60" "80" ...
#>  $ XCOV_LIANAS        : chr  "1" "2.4" "0" "6.4" ...
#>  $ H_VASC_STRATA      : chr  "1.2979" "1.117" "0.6366" "1.478" ...
#>  $ XRCOV_TALL_VEG     : chr  "50.51" "23.5" "0" "43.85" ...
#>  $ IMP_HMED_VEG       : chr  "48" "56.2" "0" "60.5" ...
#>  $ XCOV_TALL_VEG      : chr  "89" "33" "0" "67" ...
#>  $ XRCOV_SMALL_VEG    : chr  "23.27" "1.42" "70.8" "2.88" ...
#>  $ FREQ_FLOATING_AQ   : chr  "0" "0" "0" "0" ...
#>  $ XCOV_SMALL_VEG     : chr  "41" "2" "61.6" "4.4" ...
#>  $ XRCOV_MED_VEG      : chr  "14.19" "3.99" "28.51" "20.29" ...
#>  $ FREQ_SUBMERGED_AQ  : chr  "0" "0" "0" "0" ...
#>  $ XCOV_MED_VEG       : chr  "25" "5.6" "24.8" "31" ...
#>  $ J_VASC_STRATA      : chr  "0.7244" "0.6234" "0.5795" "0.8249" ...
#>  $ XRCOV_VSMALL_VEG   : chr  "2.39" "60.54" "0.69" "15.05" ...
#>  $ D_VASC_STRATA      : chr  "0.6617" "0.5684" "0.4174" "0.7224" ...
#>  $ FREQ_VTALL_VEG     : chr  "0" "0" "0" "0" ...
#>  $ IMP_VTALL_VEG      : chr  "0" "0" "0" "0" ...
#>  $ XRCOV_SUBMERGED_AQ : chr  "0" "0" "0" "0" ...
#>  $ XRCOV_FLOATING_AQ  : chr  "0" "0" "0" "0" ...
#>  $ XCOV_ARBOREAL      : chr  "0.02" "1.6" "0" "1.4" ...
#>  $ FREQ_ARBOREAL      : chr  "20" "100" "0" "40" ...
#>  $ IMP_ARBOREAL       : chr  "10.01" "50.8" "0" "20.7" ...
#>  $ N_PEAT_MOSS_DOM    : chr  "0" "1" "0" "0" ...
#>  $ XCOV_LICHENS       : chr  "0" "0" "0" "11" ...
#>  $ FREQ_LICHENS       : chr  "0" "0" "0" "100" ...
#>  $ IMP_LICHENS        : chr  "0" "0" "0" "55.5" ...
#>  $ XCOV_ALGAE         : chr  "0" "0" "0" "0" ...
#>  $ FREQ_ALGAE         : chr  "0" "0" "0" "0" ...
#>  $ IMP_ALGAE          : chr  "0" "0" "0" "0" ...
#>  $ XCOV_BRYOPHYTES    : chr  "2.02" "1.6" "0.6" "11" ...
#>  $ FREQ_BRYOPHYTES    : chr  "100" "100" "20" "100" ...
#>  $ IMP_BRYOPHYTES     : chr  "51.01" "50.8" "10.3" "55.5" ...
#>  $ FREQ_PEAT_MOSS_DOM : chr  "0" "20" "0" "0" ...
#>  $ FREQ_MACROALGAE    : chr  "0" "0" "0" "0" ...
#>  $ IMP_MACROALGAE     : chr  "0" "0" "0" "0" ...
#>  $ XCOV_MACROALGAE    : chr  "0" "0" "0" "0" ...
#>  $ MIN_H2O_DEPTH      : chr  "1" "0" "0" "0" ...
#>  $ XH2O_DEPTH_AA      : chr  "25" "2" "5.8" "0" ...
#>  $ FREQ_H2O_AQVEG     : chr  "0" "0" "0" "0" ...
#>  $ XCOV_H2O_AQVEG     : chr  "0" "0" "0" "0" ...
#>  $ MAX_COV_H2O        : chr  "100" "15" "88" "0" ...
#>  $ FREQ_H2O_NOVEG     : chr  "100" "20" "100" "0" ...
#>  $ XCOV_H2O_NOVEG     : chr  "54" "3" "2.4" "0" ...
#>  $ MAX_H2O_DEPTH      : chr  "47" "20" "30" "0" ...
#>  $ FREQ_H2O_EMERGVEG  : chr  "100" "0" "100" "0" ...
#>  $ XCOV_H2O_EMERGVEG  : chr  "13.6" "0" "41" "0" ...
#>  $ MIN_COV_H2O        : chr  "31" "0" "5" "0" ...
#>  $ FREQ_H2O           : chr  "100" "20" "100" "0" ...
#>  $ XCOV_H2O           : chr  "67.6" "3" "43.4" "0" ...
#>  $ IMP_H2O            : chr  "83.8" "11.5" "71.7" "0" ...
#>  $ XH2O_DEPTH         : chr  "25" "10" "5.8" "0" ...
#>  $ IMP_H2O_EMERGVEG   : chr  "56.8" "0" "70.5" "0" ...
#>  $ IMP_H2O_NOVEG      : chr  "77" "11.5" "51.2" "0" ...
#>  $ IMP_H2O_AQVEG      : chr  "0" "0" "0" "0" ...
#>  $ N_LITTER_TYPE      : chr  "1" "2" "1" "1" ...
#>  $ XCOV_LITTER        : chr  "21" "81" "49" "79.6" ...
#>  $ XCOV_EXPOSED_SOIL  : chr  "0" "1.6" "0.2" "14.4" ...
#>  $ FREQ_WD_FINE       : chr  "20" "100" "0" "100" ...
#>  $ MEDDEPTH_LITTER    : chr  "0" "3" "11" "3" ...
#>  $ XCOV_BAREGD        : chr  "0" "3.6" "0.2" "14.4" ...
#>  $ FREQ_WD_COARSE     : chr  "20" "80" "0" "100" ...
#>  $ XDEPTH_LITTER      : chr  "0.3" "3" "9.8" "2.9" ...
#>  $ XCOV_WD_FINE       : chr  "1.6" "10.2" "0" "12" ...
#>  $ FREQ_LITTER        : chr  "60" "100" "100" "100" ...
#>  $ LITTER_TYPE        : chr  "DECID" "DECID_FORB" "THATCH" "DECID" ...
#>  $ XCOV_WD_COARSE     : chr  "0.4" "1.8" "0" "21" ...
#>  $ XCOV_EXPOSED_GRAVEL: chr  "0" "2" "0" "0" ...
#>  $ FREQ_BAREGD        : chr  "0" "100" "20" "100" ...
#>  $ IMP_WD_COARSE      : chr  "10.2" "40.9" "0" "60.5" ...
#>   [list output truncated]
#>  - attr(*, "reshapeWide")=List of 5
#>   ..$ v.names: chr "RESULT"
#>   ..$ timevar: chr "PARAMETER"
#>   ..$ idvar  : chr "UID"
#>   ..$ times  : chr [1:108] "D_SANDT" "N_SANDT" "H_SANDT" "J_SANDT" ...
#>   ..$ varying: chr [1, 1:108] "RESULT.D_SANDT" "RESULT.N_SANDT" "RESULT.H_SANDT" "RESULT.J_SANDT" ...
head(exOut)
#>    UID D_SANDT N_SANDT H_SANDT J_SANDT DOM_SANDT FREQ_HMED_VEG N_VASC_STRATA
#> 1    1       0       1       0       0       PFO            80             6
#> 6    2       0       1       0       0       PFO           100             6
#> 11   3       0       1       0       0       PEM             0             3
#> 16   4       0       1       0       0       PFO           100             6
#> 21   5       0       1       0       0       PEM             0             3
#> 26   6    0.32       2  0.5004  0.7219       PFO           100             6
#>    IMP_SMALL_VEG XCOV_FLOATING_AQ XRCOV_VTALL_VEG IMP_MED_VEG XCOV_SUBMERGED_AQ
#> 1           70.5                0               0        62.5                 0
#> 6             41                0               0        52.8                 0
#> 11          80.8                0               0        62.4                 0
#> 16          52.2                0               0        65.5                 0
#> 21          71.5                0               0           0                 1
#> 26          60.2              4.8               0        58.9                 0
#>    XCOV_VTALL_VEG IMP_LIANAS XCOV_VSMALL_VEG FREQ_LIANAS XTOTCOV_VASC_STRATA
#> 1               0       30.5            4.22          60              176.22
#> 6               0       51.2              85         100               140.4
#> 11              0          0             0.6           0                  87
#> 16              0       53.2              23         100               152.8
#> 21              0          0             9.4           0                53.4
#> 26              0          0            40.2           0               121.2
#>    IMP_TALL_VEG XRCOV_HMED_VEG FREQ_TALL_VEG XCOV_HMED_VEG IMP_FLOATING_AQ
#> 1          94.5           9.08           100            16               0
#> 6          66.5           8.83           100          12.4               0
#> 11            0              0             0             0               0
#> 16         83.5          13.74           100            21               0
#> 21            0              0             0             0               0
#> 26           48          18.15            80            22            42.4
#>    FREQ_SMALL_VEG RG_VASC_STRATA IMP_SUBMERGED_AQ FREQ_MED_VEG XN_VASC_STRATA
#> 1             100              2                0          100            5.2
#> 6              80              1                0          100            5.8
#> 11            100              1                0          100            2.6
#> 16            100              1                0          100            5.8
#> 21            100              0             50.5            0              3
#> 26            100              1                0          100            5.6
#>    IMP_VSMALL_VEG XRCOV_LIANAS FREQ_VSMALL_VEG XCOV_LIANAS H_VASC_STRATA
#> 1           42.11         0.57              80           1        1.2979
#> 6            92.5         1.71             100         2.4         1.117
#> 11           30.3            0              60           0        0.6366
#> 16           51.5         4.19              80         6.4         1.478
#> 21           54.7            0             100           0        0.5546
#> 26           70.1            0             100           0        1.6526
#>    XRCOV_TALL_VEG IMP_HMED_VEG XCOV_TALL_VEG XRCOV_SMALL_VEG FREQ_FLOATING_AQ
#> 1           50.51           48            89           23.27                0
#> 6            23.5         56.2            33            1.42                0
#> 11              0            0             0            70.8                0
#> 16          43.85         60.5            67            2.88                0
#> 21              0            0             0           80.52                0
#> 26           13.2           61            16           16.83               80
#>    XCOV_SMALL_VEG XRCOV_MED_VEG FREQ_SUBMERGED_AQ XCOV_MED_VEG J_VASC_STRATA
#> 1              41         14.19                 0           25        0.7244
#> 6               2          3.99                 0          5.6        0.6234
#> 11           61.6         28.51                 0         24.8        0.5795
#> 16            4.4         20.29                 0           31        0.8249
#> 21             43             0               100            0        0.5048
#> 26           20.4         14.69                 0         17.8        0.9223
#>    XRCOV_VSMALL_VEG D_VASC_STRATA FREQ_VTALL_VEG IMP_VTALL_VEG
#> 1              2.39        0.6617              0             0
#> 6             60.54        0.5684              0             0
#> 11             0.69        0.4174              0             0
#> 16            15.05        0.7224              0             0
#> 21             17.6        0.3203              0             0
#> 26            33.17        0.7881              0             0
#>    XRCOV_SUBMERGED_AQ XRCOV_FLOATING_AQ XCOV_ARBOREAL FREQ_ARBOREAL
#> 1                   0                 0          0.02            20
#> 6                   0                 0           1.6           100
#> 11                  0                 0             0             0
#> 16                  0                 0           1.4            40
#> 21               1.87                 0             0             0
#> 26                  0              3.96             0             0
#>    IMP_ARBOREAL N_PEAT_MOSS_DOM XCOV_LICHENS FREQ_LICHENS IMP_LICHENS
#> 1         10.01               0            0            0           0
#> 6          50.8               1            0            0           0
#> 11            0               0            0            0           0
#> 16         20.7               0           11          100        55.5
#> 21            0               0            0            0           0
#> 26            0               0          1.2           80        40.6
#>    XCOV_ALGAE FREQ_ALGAE IMP_ALGAE XCOV_BRYOPHYTES FREQ_BRYOPHYTES
#> 1           0          0         0            2.02             100
#> 6           0          0         0             1.6             100
#> 11          0          0         0             0.6              20
#> 16          0          0         0              11             100
#> 21         62        100        81               0               0
#> 26          0          0         0             4.6              80
#>    IMP_BRYOPHYTES FREQ_PEAT_MOSS_DOM FREQ_MACROALGAE IMP_MACROALGAE
#> 1           51.01                  0               0              0
#> 6            50.8                 20               0              0
#> 11           10.3                  0               0              0
#> 16           55.5                  0               0              0
#> 21              0                  0               0              0
#> 26           42.3                  0               0              0
#>    XCOV_MACROALGAE MIN_H2O_DEPTH XH2O_DEPTH_AA FREQ_H2O_AQVEG XCOV_H2O_AQVEG
#> 1                0             1            25              0              0
#> 6                0             0             2              0              0
#> 11               0             0           5.8              0              0
#> 16               0             0             0              0              0
#> 21               0             5            10              0              0
#> 26               0             0          22.4             80            4.8
#>    MAX_COV_H2O FREQ_H2O_NOVEG XCOV_H2O_NOVEG MAX_H2O_DEPTH FREQ_H2O_EMERGVEG
#> 1          100            100             54            47               100
#> 6           15             20              3            20                 0
#> 11          88            100            2.4            30               100
#> 16           0              0              0             0                 0
#> 21           0            100             47            16               100
#> 26         100             80           25.6            67               100
#>    XCOV_H2O_EMERGVEG MIN_COV_H2O FREQ_H2O XCOV_H2O IMP_H2O XH2O_DEPTH
#> 1               13.6          31      100     67.6    83.8         25
#> 6                  0           0       20        3    11.5         10
#> 11                41           5      100     43.4    71.7        5.8
#> 16                 0           0        0        0       0          0
#> 21                53           0        0        0       0         10
#> 26              62.8          85      100       93    96.5       22.4
#>    IMP_H2O_EMERGVEG IMP_H2O_NOVEG IMP_H2O_AQVEG N_LITTER_TYPE XCOV_LITTER
#> 1              56.8            77             0             1          21
#> 6                 0          11.5             0             2          81
#> 11             70.5          51.2             0             1          49
#> 16                0             0             0             1        79.6
#> 21             76.5          73.5             0             0           0
#> 26             81.4          52.8          42.4             2          19
#>    XCOV_EXPOSED_SOIL FREQ_WD_FINE MEDDEPTH_LITTER XCOV_BAREGD FREQ_WD_COARSE
#> 1                  0           20               0           0             20
#> 6                1.6          100               3         3.6             80
#> 11               0.2            0              11         0.2              0
#> 16              14.4          100               3        14.4            100
#> 21                 0            0               0           0              0
#> 26                14          100               1          14             80
#>    XDEPTH_LITTER XCOV_WD_FINE FREQ_LITTER LITTER_TYPE XCOV_WD_COARSE
#> 1            0.3          1.6          60       DECID            0.4
#> 6              3         10.2         100  DECID_FORB            1.8
#> 11           9.8            0         100      THATCH              0
#> 16           2.9           12         100       DECID             21
#> 21             0            0           0      ABSENT              0
#> 26             1          1.2          80       DECID            6.4
#>    XCOV_EXPOSED_GRAVEL FREQ_BAREGD IMP_WD_COARSE IMP_EXPOSED_GRAVEL
#> 1                    0           0          10.2                  0
#> 6                    2         100          40.9                 11
#> 11                   0          20             0                  0
#> 16                   0         100          60.5                  0
#> 21                   0           0             0                  0
#> 26                   0          80          43.2                  0
#>    FREQ_EXPOSED_GRAVEL IMP_LITTER IMP_EXPOSED_SOIL XCOV_EXPOSED_ROCK
#> 1                    0       40.5                0                 0
#> 6                   20       90.5             50.8                 0
#> 11                   0       74.5             10.1                 0
#> 16                   0       89.8             57.2                 0
#> 21                   0          0                0                 0
#> 26                   0       49.5               47                 0
#>    IMP_WD_FINE IMP_BAREGD IMP_EXPOSED_ROCK FREQ_EXPOSED_SOIL FREQ_EXPOSED_ROCK
#> 1         10.8          0                0                 0                 0
#> 6         55.1       51.8                0               100                 0
#> 11           0       10.1                0                20                 0
#> 16          56       57.2                0               100                 0
#> 21           0          0                0                 0                 0
#> 26        50.6         47                0                80                 0