Calculate Wetland Indicator Status metrics

This function calculates Wetland Indicator Status (WIS) metrics, including variations based on native status, if NWCA_NATSTAT is present in the input data frame.

calcWIS(vascIn, sampID = "UID")

Arguments

vascIn

Data frame containing cover data summarized by UID and TAXON, with the following fields:

sampID: Variable(s) identified in sampID argument

TAXON: Taxon name

WIS: Wetland Indicator Status, with possible values of FAC, FACU, FACW, OBL, UPL, NL, TBD, or missing. NL is recoded to UPL in function, TBD is recoded to missing in function.

FREQ: Frequency of taxon among plots at site

XABCOV: Mean percent cover of taxon across plots

TOTN: Number of taxa in sample

sXRCOV: proportion of summed cover across all taxa (XTOTABCOV) represented by taxon in sample

NWCA_NATSTAT (optional): Native status variable with categories of 'NAT', 'ADV', 'CRYP', 'INTR', 'UND'. UND taxa are ignored.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

Value

Data frame containing sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait value (represented as TRAITNM below):

N_TRAITNM: Number of taxa with trait

PCTN_TRAITNM: Number of taxa with trait as percentage of TOTN

XABCOV_TRAITNM: Sum of XABCOV values across taxa with trait

XRCOV_TRAITNM: Sum of sXRCOV values across taxa with trait

In addition WIS indices based on all and native species only (if NWCA_NATSTAT is provided in the input data frame), with the suffixes ALL and NAT, respectively. WIS values recoded as numeric with OBL=1, FACW=2, FAC=3, FACU=4, UPL=5 for WETIND metrics and as OBL=5, FACW=4, FAC=3 FACU=2, and UPD=1 for WETIND2 metrics:

WETIND_COV_ALL, WETIND_COV_NAT: Wetland Index, weighted by taxon cover, based on numeric conversion of WIS, with lower numbers indicating wetter conditions

WETIND_FREQ_ALL, WETIND_FREQ_NAT: Wetland Index, weight by frequency, based on numeric conversion of WIS, with lower numbers indicating wetter conditions

WETIND2_COV_ALL, WETIND2_COV_NAT: Wetland Index, weighted by taxon cover, based on numeric conversion of WIS, with higher numbers indicating wetter conditions

WETIND2_FREQ_ALL, WETIND2_FREQ_NAT: Wetland Index, weight by frequency, based on numeric conversion of WIS, with higher numbers indicating wetter conditions

If NWCA_NATSTAT is provided in the input data frame, the following metrics are calculated based on Alien + Cryptogenic species:

N_OBLFACW_AC: Number of alien and cryptogenic taxa with WIS of either OBL or FACW.

XABCOV_OBLFACW_AC: Mean absolute cover of alien and cryptogenic taxa with WIS of either OBL or FACW.

XRCOV_OBLFACW_AC: Mean relative cover of alien and cryptogenic taxa with WIS of either OBL or FACW.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom Blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
exPlant <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME",
  inTaxa = taxaNWCA, inNat = ccNatNWCA, inCVal = ccNatNWCA,
  inWIS = wisNWCA, cValReg = "STATE"
)
#> [1] "Will not be able to calculate metrics using DUR_ALT,GRH_ALT,HERB,TREE_COMB,SHRUB_COMB,VINE_ALL without these parameters in inTaxa"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "byUID"  "byPlot"
#>  [1] "TAXON"           "STATE"           "UID"             "USAC_REGION"    
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"        
#>  [1] "TAXON"           "USAC_REGION"     "STATE"           "UID"            
#>  [5] "NUM"             "XABCOV"          "NPLOTS"          "DISTINCT"       
#>  [9] "SPECIES_NAME_ID" "FAMILY"          "GENUS"           "CATEGORY"       
#> [13] "GROWTH_HABIT"    "DURATION"        "NWCA_CC"         "NWCA_NATSTAT"   
#> [17] "WIS"            
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

wisEx <- calcWIS(exPlant$byUIDspp)

head(wisEx)
#>          UID PARAMETER RESULT
#> 5.N_FAC    5     N_FAC      0
#> 8.N_FAC    8     N_FAC      0
#> 10.N_FAC  10     N_FAC      0
#> 7.N_FAC    7     N_FAC      1
#> 9.N_FAC    9     N_FAC      3
#> 3.N_FAC    3     N_FAC      4
unique(wisEx$PARAMETER)
#>  [1] "N_FAC"               "N_FACU"              "N_FACW"             
#>  [4] "N_OBL"               "N_UPL"               "PCTN_FAC"           
#>  [7] "PCTN_FACU"           "PCTN_FACW"           "PCTN_OBL"           
#> [10] "PCTN_UPL"            "XABCOV_FAC"          "XABCOV_FACU"        
#> [13] "XABCOV_FACW"         "XABCOV_OBL"          "XABCOV_UPL"         
#> [16] "XRCOV_FAC"           "XRCOV_FACU"          "XRCOV_FACW"         
#> [19] "XRCOV_OBL"           "XRCOV_UPL"           "N_OBL_FACW"         
#> [22] "PCTN_OBL_FACW"       "XABCOV_OBL_FACW"     "XRCOV_OBL_FACW"     
#> [25] "N_OBL_FACW_FAC"      "PCTN_OBL_FACW_FAC"   "XABCOV_OBL_FACW_FAC"
#> [28] "XRCOV_OBL_FACW_FAC"  "N_FAC_FACU"          "PCTN_FAC_FACU"      
#> [31] "XABCOV_FAC_FACU"     "XRCOV_FAC_FACU"      "WETIND_COV_ALL"     
#> [34] "WETIND_FREQ_ALL"     "WETIND2_COV_ALL"     "WETIND2_FREQ_ALL"   
#> [37] "WETIND_COV_NAT"      "WETIND_FREQ_NAT"     "WETIND2_COV_NAT"    
#> [40] "WETIND2_FREQ_NAT"    "N_OBLFACW_AC"        "XRCOV_OBLFACW_AC"   
#> [43] "XABCOV_OBLFACW_AC"