Calculate water cover metrics

This function calculates metrics water cover and depth based on ground cover data. This function is called by calcVtype_GcovMets().

calcWcovMets(dataIn, nPlot, sampID = "UID", survyear = "2011")

Arguments

dataIn

A data frame containing the following variables:

sampID: Variable(s) identified in sampID argument

PLOT: Sample plot from which data were collected

PARAMETER: specific measurement type

RESULT: measured value

The following parameters are used in calculating water cover metrics: 'TIME', 'MINIMUM_DEPTH', 'MAXIMUM_DEPTH', 'PREDOMINANT_DEPTH', 'TOTAL_WATER', 'WATER_NOVEG', 'WATER_AQVEG', 'WATER_EMERGVEG'. Additional parameters or variables are ignored. WATER_NOVEG, WATER_AQVEG, MINIMUM_DEPTH, and MAXIMUM_DEPTH not measured in 2016.

nPlot

A data frame with the number of plots sampled associated with each sample with sampID variables and NPLOTS

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default.

survyear

A string with the survey year. The default is '2011'. Starting in 2016, only TIME, TOTAL_WATER, and PREDOMINANT_DEPTH is measured.

Value

Either a character string containing an error message when metric calculation is not successful, or a data frame. Data frame contains sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait value, with valid values of: MIN_H2O_DEPTH, MAX_H2O_DEPTH, XH2O_DEPTH_AA, MIN_COV_H2O, MAX_COV_H2O, FREQ_H2O, FREQ_H2O_AQVEG, FREQ_H2O_EMERGVEG, FREQ_H2O_NOVEG, XCOV_H2O, XCOV_H2O_AQVEG, XCOV_H2O_EMERGVEG, XCOV_H2O_NOVEG, IMP_H2O, IMP_H2O_AQVEG, IMP_H2O_EMERGVEG, IMP_H2O_NOVEG, XH2O_DEPTH. A list of metric descriptions is provided in the document named VegTypes_GrdCover_Metric_Descriptions.pdf included in the help directory for the package.

Details

If any of the parameters are missing, they are assumed to be zeros (if numeric), and metric values associated with any metrics that cannot be calculated due to missing parameters are set to a standardized value.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(Vtype_GrCovEx)
#>   UID PLOT       PARAMETER RESULT
#> 1   7    4   MINIMUM_DEPTH      8
#> 2   7    3 LITTER_DEPTH_NE      0
#> 3   7    5        TALL_VEG      0
#> 4   7    4     WATER_AQVEG      0
#> 5   7    5   MINIMUM_DEPTH      8
#> 6   7    5     WATER_NOVEG      0
# Create data frame with number of plots sampled for each sampID
nplots <- data.frame(UID = seq(1:10), NPLOTS = rep(5, 10), stringsAsFactors = F)
# alternative approach to creating this data frame
nplots <- aggregate(
  x = list(NPLOTS = Vtype_GrCovEx$PLOT),
  by = Vtype_GrCovEx[c("UID")],
  FUN = function(x) {
    length(unique(x))
  }
)

wcovEx <- calcWcovMets(Vtype_GrCovEx, nplots)
#> [1] "Done with water depth metrics"

head(wcovEx)
#>   UID      PARAMETER RESULT
#> 1   1  MIN_H2O_DEPTH      1
#> 2   1  XH2O_DEPTH_AA     25
#> 3   1 FREQ_H2O_AQVEG      0
#> 4   1 XCOV_H2O_AQVEG      0
#> 5   1    MAX_COV_H2O    100
#> 6   1 FREQ_H2O_NOVEG    100
unique(wcovEx$METRIC)
#> NULL