Create data frames for function input

This internal function merges the inputs of plant cover data and taxalist and summarizes by the taxonomic level specified. Not intended for use on its own.

createDFs(
  sampID = "UID",
  tvar,
  vascIn,
  taxon_name,
  taxa,
  state = "STATE",
  coeReg = "USAC_REGION",
  cValReg = "NWC_CREG"
)

Arguments

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

tvar

String with the level of taxonomy (taxon_name,'GENUS','FAMILY')

vascIn

Data frame with vegetation cover data, having the following variables:

sampID: Variable(s) found in the argument sampID

PLOT: Plot number of data (1 to 5 possible)

Variable named in state: Two-letter state postal code of site, used to link native status to taxa in native status taxalist (inNat)

Variable named in coeReg: U.S. Army Corps of Engineers region abbreviation for sample, to correspond to GEOG_ID in Wetland Indicator Status taxalist (inWIS)

Variable named in cValReg: NWCA C-value region abbreviation for sample, to correspond those in C-value taxalist

Variable named in taxon_name: Taxon name, must match with taxa data frame

COVER: Percentage estimated vegetation cover by TAXON in PLOT

taxon_name

String containing the name of variable for taxon name in vascIn and in taxalists.

taxa

Data frame containing USDA_NAME, GENUS, CATEGORY, GROWTH_HABIT, and DURATION variables. Dataset taxaNWCA is the default if not specified. These variables are assumed to be populated with abbreviations as found in USDA PLANTS database.

state

String containing the name of the state in vascIn, with default value of 'STATE'

coeReg

String containing the name of the U.S. Army Corps of Engineers region in vascIn associated with Wetland Indicator Status, with default value of 'USAC_REGION'

cValReg

String containing the name of the variable in vascIn which specifies the C-value region.

Value

A list containing two data frames. The first data frame summarizes data by sampID variables and TAXON and contains sampID variables, STATE, USAC_REGION, TAXON, NUM, XABCOV, DISTINCT, and NPLOTS. NUM is the number of plots in which taxon occurs, and XABCOV is the mean absolute COVER across plots. DISTINCT is the value 1 assigned to each row. NPLOTS is the number of plots sampled.

The second summarizes by sampID variables, PLOT, and TAXON. Each data frame contains sampID variables, PLOT, STATE, USAC_REGION, TAXON, COVER, and DISTINCT. DISTINCT assigns the value for each taxon as 1 if the taxon has COVER>0 or 0 if not. COVER is the sum of the COVER variable by sampID variables, PLOT, and TAXON.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
data(taxaNWCA)

outEx <- createDFs(
  sampID = "UID", "GENUS", VascPlantEx, taxon_name = "USDA_NAME",
  taxaNWCA, cValReg = "STATE"
)
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
head(outEx$byUID)
#>   UID STATE USAC_REGION        TAXON NUM XABCOV NPLOTS DISTINCT
#> 1   1    DE      E_MTNS         ACER   5  57.00      5        1
#> 2   1    DE      E_MTNS        APIOS   1   0.02      5        1
#> 3   1    DE      E_MTNS        CAREX   1   0.02      5        1
#> 4   1    DE      E_MTNS CEPHALANTHUS   3   3.04      5        1
#> 5   1    DE      E_MTNS      CLETHRA   3   4.40      5        1
#> 6   1    DE      E_MTNS    DIOSPYROS   1   0.20      5        1
head(outEx$byPlot)
#>   UID STATE USAC_REGION PLOT TAXON COVER DISTINCT
#> 1   4    GA  CSTL_PLAIN    1  ACER    10        1
#> 2   2    MD  CSTL_PLAIN    1  ACER    10        1
#> 3   1    DE      E_MTNS    1  ACER    50        1
#> 4   6    ND   GT_PLAINS    1  ACER     4        1
#> 5   4    GA  CSTL_PLAIN    2  ACER     3        1
#> 6   2    MD  CSTL_PLAIN    2  ACER     0        0