NWCA 2016 Coefficient of Conservation Values (C-values)

A data frame containing NWCA 2016 state- or region-specific Coefficient of Conservatism values for vascular plant taxa

cvalNWCA_2016

Format

A data frame containing 24206 observations on the following variables

SPECIES_NAME_ID

A numeric vector containing the NWCA taxonomic ID number

NWCA_NAME

Accepted USDA PLANTS name, or the official NWCA name where not available

GEOG_TYPE

Type of Geographic Unit

GEOG_ID

Geographic unit to which value applies, which is NWC_CREG16 in this dataset

NWCA_CVAL

C-value region-specific Coefficient of Conservatism value as used in NWCA 2016

Note

This dataset is the list of C-values used for site-specific and taxon-specific assignments for NWCA 2016.

Examples

head(cvalNWCA_2016)
#>   SPECIES_NAME_ID            NWCA_NAME  GEOG_TYPE  GEOG_ID NWCA_CVAL
#> 1              58 xELYHORDEUM MACOUNII NWC_CREG16       ND         3
#> 2             129                ABIES NWC_CREG16    MD_AP         9
#> 3             135       ABIES BALSAMEA NWC_CREG16 MA_NAAMH         4
#> 4             135       ABIES BALSAMEA NWC_CREG16 ME_NAAMH         4
#> 5             135       ABIES BALSAMEA NWC_CREG16  ME_NBPH         3
#> 6             135       ABIES BALSAMEA NWC_CREG16   ME_NCZ         6
str(cvalNWCA_2016)
#> 'data.frame':	24206 obs. of  5 variables:
#>  $ SPECIES_NAME_ID: int  58 129 135 135 135 135 135 135 135 135 ...
#>  $ NWCA_NAME      : chr  "xELYHORDEUM MACOUNII" "ABIES" "ABIES BALSAMEA" "ABIES BALSAMEA" ...
#>  $ GEOG_TYPE      : chr  "NWC_CREG16" "NWC_CREG16" "NWC_CREG16" "NWC_CREG16" ...
#>  $ GEOG_ID        : chr  "ND" "MD_AP" "MA_NAAMH" "ME_NAAMH" ...
#>  $ NWCA_CVAL      : chr  "3" "9" "4" "4" ...