Calculate vascular plant richness metrics

This internal function calculates taxa richness of sample at a specified level of taxonomy, based on output from CreateDFs(). This function used in calcRichness() function.

int.calcRich(x, y, tlevel, sampID)

Arguments

x

Data frame containing cover data summarized by sampID variables, TAXON, DISTINCT at the specified taxonomic level (tlevel)

y

Data frame containing cover data summarized by sampID variables, PLOT, TAXON, and DISTINCT at the specified taxonomic level.

tlevel

Taxonomic level of input dataset, abbreviated to serve as suffix to metric names ('SPP','GEN','FAM')

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples

Value

Data frame containing sampID variables, PARAMETER, and RESULT, with one row of results per parameter and sample. The values for PARAMETER consist of the metric name concatenated with taxonomic level (represented as TAXLEVEL below):

TOTN_TAXLEVEL: Number of unique taxa in sample

XN_TAXLEVEL: Mean number of taxa per plot

MEDN_TAXLEVEL: Median number of taxa per plot

SDN_TAXLEVEL: Standard deviation of number of taxa per plot

N_PLOTS: Number of plots sampled for sampID

Warning

This function not intended for use on its own

Author

Karen Blocksom