Calculate metrics using traits with only two values (0/1)

This internal function calculates metrics for traits that are indicator values (0 or 1), specifically for alien and cryptogenic species groups. Used by calcNative() function.

int.calcTraits_Indicator.alt(vascIn, trait, sampID)

Arguments

vascIn

Input data frame with sampID variables, TAXON, TOTN, XABCOV, sXRCOV, sRFREQ, and variable with name in argument trait. TOTN is the total number of taxa at the lowest level for each sample (combination of sampID variables). XABCOV is the mean absolute cover of taxon. sXRCOV is the percentage of the total sum of absolute cover across all taxa for a sample. sRFREQ is the relative frequency of a taxon, calculated as the percentage of the total frequency of taxon occurrence across all taxa for a sample.

trait

Character string containing name of binary trait variable.

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples

Value

Data frame containing sampID variables, PARAMETER, RESULT, where values of PARAMETER consist of the metric name concatenated with trait name (represented as TRAITNM below):

N_TRAITNM: Number of taxa with trait value

PCTN_TRAITNM: Number of taxa with trait value as percentage of TOTN

XABCOV_TRAITNM: Sum of XABCOV values across taxa with trait value

XRCOV_TRAITNM: Sum of sXRCOV values across taxa with trait value

RFREQ_TRAITNM: Sum of sRFREQ values across taxa with trait value

RIMP_TRAITNM: Relative importance ((RFREQ_TRAITVAL + XRCOV_TRAITVAL)/2) of taxa with trait value

Warning

This function not intended for use on its own

Author

Karen Blocksom