Prepare data for calculating metrics - alternative version

Assemble datasets from data frame containing vegetation cover by sampID variables and PLOT, and taxa lists containing variables used in metric calculation. Species-, genus-, and family-level datasets are summarized by plot and sample. For the species-level output data summarized by sampID variables, various traits are added to the output data set.

nwcaVegData(
  vascIn,
  sampID = "UID",
  taxon_name,
  inTaxa = taxaNWCA,
  inNat = ccNatNWCA,
  inCVal = ccNatNWCA,
  inWIS = wisNWCA,
  state = "STATE",
  coeReg = "USAC_REGION",
  cValReg = "STATE"
)

Arguments

vascIn

Data frame containing cover data summarized by sampID variables, PLOT, and taxon value, with the following variable names:

sampID A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

PLOT: Plot number of data

variable name provided in taxon_name: Taxon name, based on USDA PLANTS names, supplemented with NWCA names, if not in USDA PLANTS

COVER: Percent cover of taxon in plot, including zeros for plots in which a taxon does not occur.

Variable named in state: Two-letter state postal code of site, used to link native status to taxa in native status taxalist (inNat)

Variable named in coeReg: U.S. Army Corps of Engineers region abbreviation for sample, to correspond to GEOG_ID in Wetland Indicator Status taxalist (inWIS)

Variable named in cValReg: NWCA C-value regions: values must match GEOG_ID in C-value taxalist (inCVal)

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

taxon_name

String containing the name of variable for taxon name in vascIn and in taxalists.

inTaxa

Data frame with all taxa in vascIn, with variables:

variable name provided in taxon_name: Taxon name, based on USDA PLANTS names, supplemented with NWCA names, if not in USDA PLANTS

FAMILY: Family name of taxon

GENUS: Genus name of taxon

CATEGORY (optional): USDA PLANTS category variable, necessary to calculate category metrics.

DURATION (optional): USDA PLANTS duration variable, necessary to calculate duration metrics.

GROWTH_HABIT (optional): USDA PLANTS growth habit variable,necessary to calculate growth habit metrics

SPECIES_NAME_ID (optional): Taxonomic ID number, which will be used in Bray-Curtis distance metrics if available.

inNat

Data frame with native status:

variable name provided in taxon_name: Taxon name, based on USDA PLANTS names, supplemented with NWCA names, if not in USDA PLANTS

GEOG_ID: Postal abbreviation for STATE of taxon

NWCA_NATSTAT: Native status variable, as used in NWCA, necessary to calculate native status metrics.

inCVal

Data frame with coefficient of conservatism values:

USDA_NAME: Taxon name

GEOG_ID: Code indicating C region for site, as supplied with cover data.

NWCA_CVAL: Coefficient of conservatism (C-value), as used in NWCA, necessary to calculate metrics based on C-values.

inWIS

Data frame with Wetland Indicator Status, from U.S. Army Corps of Engineers (USAC):

variable name provided in taxon_name: Taxon name, based on USDA PLANTS names, supplemented with NWCA names, if not in USDA PLANTS

GEOG_ID: USAC region, abbreviated to match those used in input data frame

WIS: Wetland Indicator Status as provided by USAC or added for NWCA

state

String containing the name of the state in vascIn, with default value of 'STATE'

coeReg

String containing the name of the U.S. Army Corps of Engineers region in vascIn associated with Wetland Indicator Status, with default value of 'USAC_REGION'

cValReg

String containing the name of the Coefficient of Conservatism region in vascIn, with default value of 'STATE'

Value

A list containing six data frames:

byUIDspp: Data frame with data summarized by sampID variables and TAXON at the species level and contains:

sampID: Variable(s) identified in sampID argument

cValReg (as specified in function arguments): Coefficient of Conservatism region associated with site, necessary for merging with inCVal, with default of 'STATE'

state: Two-letter state code value (will match name specified in function arguments)

coeReg: U.S. Army Corps of Engineers region code value (will match name specified in function arguments)

TAXON: Taxon name

NUM: Number of occurrences of taxon across plots

XABCOV: Mean percent absolute cover of taxon across plots

DISTINCT: Distinctness value for each taxon, 1 if the taxon has COVER>0 and 0 if not.

NPLOTS: Number of plots in sample (1-5)

TOTN: Total number of taxa in sample

XTOTABCOV: Sum of XABCOV across all taxa in sample

sXRCOV: taxon mean relative cover (XABCOV/XTOTABCOV)*100

FREQ: Relative number of plots in which taxon occurs (NUM/NPLOTS)*100

TOTFREQ: Sum of FREQ across all taxa in sample

SRFREQ: Relative frequency of taxon relative to total frequency (FREQ/TOTFREQ)*100

This data frame is also merged with the input taxa data frames and contains, in addition, GENUS and FAMILY, but also contains (depending on the input taxa lists): CATEGORY, DURATION, GROWTH_HABIT, NWCA_CC, NWCA_NATSTAT, WIS.

byPlotspp: Data frame with data summarized by sampID variables, PLOT, and TAXON at the species level. Each data frame contains:

sampID Variables identified in sampID argument

PLOT: Plot number

state: Two-letter state code value associated with native status (will match name specified in function arguments)

coeReg: U.S. Army Corps of Engineers region code value (will match name specified in function arguments)

cValReg: Site region associated with Coefficient of Conservatism value (will match name specified in function arguments)

TAXON: Taxon name

COVER: Sum of cover by TAXON within plot

DISTINCT: Distinctness of taxon, value of 1 assigned to each row

byUIDgen: Data frame with data summarized by sampID variables and TAXON at the genus level and contains sampID, cValReg variable, coeReg, TAXON, NUM, XABCOV, and DISTINCT. NUM is the number of plots in which taxon occurs, and XABCOV is the mean absolute COVER across plots. DISTINCT is the value 1 assigned to each row.

byPlotgen: Data frame with data summarized by sampID variables, PLOT, and TAXON at the genus level. Each data frame contains sampID, PLOT, cValReg variable, coeReg, TAXON, COVER, and DISTINCT. DISTINCT assigns the value for each taxon as 1 if the taxon has COVER>0 or 0 if not. COVER is the sum of the COVER variable.

byUIDfam: Data frame with data summarized by sampID variables and TAXON at the family level and contains sampID, cValReg variable, coeReg, TAXON, NUM, XABCOV, and DISTINCT. NUM is the number of plots in which taxon occurs, and XABCOV is the mean absolute COVER across plots. DISTINCT is the value 1 assigned to each row.

byPlotfam: Data frame with data summarized by sampID , PLOT, and TAXON at the family level. Each data frame contains sampID, PLOT, cValReg variable, coeReg, TAXON, COVER, and DISTINCT. DISTINCT assigns the value for each taxon as 1 if the taxon has COVER>0 or 0 if not. COVER is the sum of the COVER variable.

Details

This function calls the createDFs() function, which sums cover by sampID variables, PLOT, TAXON, with sums > 100 truncated to 100 percent.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
prepEx <- nwcaVegData(vascIn = VascPlantEx, taxon_name = "USDA_NAME", cValReg = "STATE")
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done with plot summing"
#> [1] "Done with sampID summing"
#> [1] "Done preparing datasets"

str(prepEx)
#> List of 6
#>  $ byUIDspp :'data.frame':	238 obs. of  30 variables:
#>   ..$ UID              : int [1:238] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ TAXON            : chr [1:238] "ACER RUBRUM" "DIOSPYROS VIRGINIANA" "MAGNOLIA VIRGINIANA" "LIRIODENDRON TULIPIFERA" ...
#>   ..$ USAC_REGION      : chr [1:238] "E_MTNS" "E_MTNS" "E_MTNS" "E_MTNS" ...
#>   ..$ STATE            : chr [1:238] "DE" "DE" "DE" "DE" ...
#>   ..$ NUM              : num [1:238] 5 1 1 1 4 1 5 1 5 4 ...
#>   ..$ XABCOV           : num [1:238] 57 0.2 0.2 0.02 24.42 ...
#>   ..$ NPLOTS           : int [1:238] 5 5 5 5 5 5 5 5 5 5 ...
#>   ..$ DISTINCT         : num [1:238] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ SPECIES_NAME_ID.x: int [1:238] 866 28486 52662 49686 57631 5290 87875 71431 34059 72682 ...
#>   ..$ DIVISION         : chr [1:238] "MAGNOLIOPHYTA" "MAGNOLIOPHYTA" "MAGNOLIOPHYTA" "MAGNOLIOPHYTA" ...
#>   ..$ ORDER            : chr [1:238] "SAPINDALES" "EBENALES" "MAGNOLIALES" "MAGNOLIALES" ...
#>   ..$ FAMILY           : chr [1:238] "ACERACEAE" "EBENACEAE" "MAGNOLIACEAE" "MAGNOLIACEAE" ...
#>   ..$ GENUS            : chr [1:238] "ACER" "DIOSPYROS" "MAGNOLIA" "LIRIODENDRON" ...
#>   ..$ SPECIES          : chr [1:238] "RUBRUM" "VIRGINIANA" "VIRGINIANA" "TULIPIFERA" ...
#>   ..$ SUBSPECIES       : chr [1:238] "" "" "" "" ...
#>   ..$ VARIETY          : chr [1:238] "" "" "" "" ...
#>   ..$ CATEGORY         : chr [1:238] "DICOT" "DICOT" "DICOT" "DICOT" ...
#>   ..$ GROWTH_HABIT     : chr [1:238] "TREE" "TREE" "TREE, SHRUB" "TREE" ...
#>   ..$ DURATION         : chr [1:238] "PERENNIAL" "PERENNIAL" "PERENNIAL" "PERENNIAL" ...
#>   ..$ SPECIES_NAME_ID.y: int [1:238] 866 28486 52662 49686 57631 5290 87875 71431 34059 72682 ...
#>   ..$ NWCA_CC          : chr [1:238] "2" "4" "6" "4" ...
#>   ..$ NWCA_NATSTAT     : chr [1:238] "NAT" "NAT" "NAT" "NAT" ...
#>   ..$ SPECIES_NAME_ID  : int [1:238] 866 28486 52662 49686 57631 5290 87875 71431 34059 72682 ...
#>   ..$ WIS              : chr [1:238] "FAC" "FAC" "FACW" "FACU" ...
#>   ..$ XTOTABCOV        : num [1:238] 204 204 204 204 204 ...
#>   ..$ TOTFREQ          : num [1:238] 920 920 920 920 920 920 920 920 920 920 ...
#>   ..$ TOTN             : int [1:238] 18 18 18 18 18 18 18 18 18 18 ...
#>   ..$ sXRCOV           : num [1:238] 28.00157 0.09825 0.09825 0.00983 11.99646 ...
#>   ..$ FREQ             : num [1:238] 100 20 20 20 80 20 100 20 100 80 ...
#>   ..$ sRFREQ           : num [1:238] 10.87 2.17 2.17 2.17 8.7 ...
#>  $ byPlotspp:'data.frame':	1190 obs. of  10 variables:
#>   ..$ STATE          : chr [1:1190] "DE" "DE" "DE" "DE" ...
#>   ..$ TAXON          : chr [1:1190] "ACER RUBRUM" "ACER RUBRUM" "ACER RUBRUM" "ACER RUBRUM" ...
#>   ..$ UID            : int [1:1190] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ USAC_REGION    : chr [1:1190] "E_MTNS" "E_MTNS" "E_MTNS" "E_MTNS" ...
#>   ..$ PLOT           : int [1:1190] 1 2 3 4 5 1 2 3 4 5 ...
#>   ..$ COVER          : num [1:1190] 50 80 90 50 15 0 0 0.1 0 0 ...
#>   ..$ DISTINCT       : num [1:1190] 1 1 1 1 1 0 0 1 0 0 ...
#>   ..$ SPECIES_NAME_ID: int [1:1190] 866 866 866 866 866 5290 5290 5290 5290 5290 ...
#>   ..$ NWCA_CC        : chr [1:1190] "2" "2" "2" "2" ...
#>   ..$ NWCA_NATSTAT   : chr [1:1190] "NAT" "NAT" "NAT" "NAT" ...
#>  $ byUIDgen :'data.frame':	204 obs. of  8 variables:
#>   ..$ UID        : int [1:204] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ STATE      : chr [1:204] "DE" "DE" "DE" "DE" ...
#>   ..$ USAC_REGION: chr [1:204] "E_MTNS" "E_MTNS" "E_MTNS" "E_MTNS" ...
#>   ..$ TAXON      : chr [1:204] "ACER" "APIOS" "CAREX" "CEPHALANTHUS" ...
#>   ..$ NUM        : num [1:204] 5 1 1 3 3 1 5 1 1 5 ...
#>   ..$ XABCOV     : num [1:204] 57 0.02 0.02 3.04 4.4 ...
#>   ..$ NPLOTS     : int [1:204] 5 5 5 5 5 5 5 5 5 5 ...
#>   ..$ DISTINCT   : num [1:204] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ byPlotgen:'data.frame':	1020 obs. of  7 variables:
#>   ..$ UID        : int [1:1020] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ STATE      : chr [1:1020] "DE" "DE" "DE" "DE" ...
#>   ..$ USAC_REGION: chr [1:1020] "E_MTNS" "E_MTNS" "E_MTNS" "E_MTNS" ...
#>   ..$ TAXON      : chr [1:1020] "ACER" "ACER" "ACER" "ACER" ...
#>   ..$ PLOT       : int [1:1020] 1 2 3 4 5 1 2 3 4 5 ...
#>   ..$ COVER      : num [1:1020] 50 80 90 50 15 0 0 0.1 0 0 ...
#>   ..$ DISTINCT   : num [1:1020] 1 1 1 1 1 0 0 1 0 0 ...
#>  $ byUIDfam :'data.frame':	147 obs. of  8 variables:
#>   ..$ UID        : int [1:147] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ STATE      : chr [1:147] "DE" "DE" "DE" "DE" ...
#>   ..$ USAC_REGION: chr [1:147] "E_MTNS" "E_MTNS" "E_MTNS" "E_MTNS" ...
#>   ..$ TAXON      : chr [1:147] "ACERACEAE" "AQUIFOLIACEAE" "CLETHRACEAE" "CORNACEAE" ...
#>   ..$ NUM        : num [1:147] 5 1 3 4 1 1 5 1 1 5 ...
#>   ..$ XABCOV     : num [1:147] 57 3 4.4 24.42 0.02 ...
#>   ..$ NPLOTS     : int [1:147] 5 5 5 5 5 5 5 5 5 5 ...
#>   ..$ DISTINCT   : num [1:147] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ byPlotfam:'data.frame':	735 obs. of  7 variables:
#>   ..$ UID        : int [1:735] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ STATE      : chr [1:735] "DE" "DE" "DE" "DE" ...
#>   ..$ USAC_REGION: chr [1:735] "E_MTNS" "E_MTNS" "E_MTNS" "E_MTNS" ...
#>   ..$ TAXON      : chr [1:735] "ACERACEAE" "ACERACEAE" "ACERACEAE" "ACERACEAE" ...
#>   ..$ PLOT       : int [1:735] 1 2 3 4 5 1 2 3 4 5 ...
#>   ..$ COVER      : num [1:735] 50 80 90 50 15 15 0 0 0 0 ...
#>   ..$ DISTINCT   : num [1:735] 1 1 1 1 1 1 0 0 0 0 ...