Prepare data for calculating metrics

Assemble datasets from data frame containing vegetation cover by sampID variables and PLOT, and taxa lists containing variables used in metric calculation. Species-, genus-, and family-level datasets are summarized by plot and sample. For the species-level output data summarized by sampID variables, various traits are added to the output data set.

prepareData(
  vascIn,
  sampID = "UID",
  taxon_name,
  inTaxa = taxaNWCA_2016,
  inNat = nativeNWCA_2016,
  inCVal = cvalNWCA_2016,
  inWIS = wisNWCA_2016,
  state = "STATE",
  coeReg = "USAC_REGION",
  cValReg = "NWC_CREG"
)

Arguments

vascIn

Data frame containing cover data summarized by sampID variables, PLOT, and taxon value, with the following variable names:

sampID A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

PLOT: Plot number of data

variable name provided in taxon_name: Taxon name, based on USDA PLANTS names, supplemented with NWCA names, if not in USDA PLANTS

COVER: Percent cover of taxon in plot, including zeros for plots in which a taxon does not occur.

Variable named in state: Two-letter state postal code of site, used to link native status to taxa in native status taxalist (inNat)

Variable named in coeReg: U.S. Army Corps of Engineers region abbreviation for sample, to correspond to GEOG_ID in Wetland Indicator Status taxalist (inWIS)

Variable named in cValReg: NWCA C-value regions: values must match GEOG_ID in C-value taxalist (inCVal)

sampID

A character vector containing the name(s) of variable(s) necessary to identify unique samples, 'UID' by default

taxon_name

String containing the name of variable for taxon name in vascIn and in taxalists.

inTaxa

Data frame with all taxa in vascIn, with variables:

variable name provided in taxon_name: Taxon name, consistent with name in vascIn data frame.

FAMILY: Family name of taxon

GENUS: Genus name of taxon

CATEGORY (optional): USDA PLANTS category variable, necessary to calculate category metrics.

DURATION (optional): USDA PLANTS duration variable, necessary to calculate duration metrics.

GROWTH_HABIT (optional): USDA PLANTS growth habit variable,necessary to calculate growth habit metrics

SPECIES_NAME_ID (optional): Taxonomic ID number, which will be used in Bray-Curtis distance metrics if available.

inNat

Data frame with native status:

variable name provided in taxon_name: Taxon name, consistent with name in vascIn data frame.

GEOG_ID: Postal abbreviation for STATE of taxon

NWCA_NATSTAT: Native status variable, as used in NWCA, necessary to calculate native status metrics.

inCVal

Data frame with coefficient of conservatism values:

variable name provided in taxon_name: Taxon name, consistent with name in vascIn data frame.

GEOG_ID: Code indicating C region for site, as supplied with cover data.

NWCA_CC: Coefficient of conservatism (C-value), as used in NWCA, necessary to calculate metrics based on C-values.

inWIS

Data frame with Wetland Indicator Status, from U.S. Army Corps of Engineers (USAC):

variable name provided in taxon_name: Taxon name, consistent with name in vascIn data frame.

GEOG_ID: USAC region, abbreviated to match USAC_REGION used in input data frame

WIS: Wetland Indicator Status as provided by USAC or added for NWCA

state

String containing the name of the state in vascIn, with default value of 'STATE'

coeReg

String containing the name of the U.S. Army Corps of Engineers region in vascIn associated with Wetland Indicator Status, with default value of 'USAC_REGION'

cValReg

String containing the name of the variable in vascIn which specifies the C-value region.

Value

A list containing six data frames:

byUIDspp: Data frame with data summarized by sampID variables and TAXON at the species level and contains:

sampID: Variable(s) identified in sampID argument

Variable named in state: Two-letter state postal code of site, used to link native status to taxa in native status taxalist (inNat)

Variable named in coeReg: U.S. Army Corps of Engineers region abbreviation for sample, to correspond to GEOG_ID in Wetland Indicator Status taxalist (inWIS)

Variable named in cValReg: NWCA C-value regions: values must match GEOG_ID in C-value taxalist (inCVal)

TAXON: Taxon name

NUM: Number of occurrences of taxon across plots

XABCOV: Mean percent absolute cover of taxon across plots

DISTINCT: Distinctness value for each taxon, 1 if the taxon has COVER>0 and 0 if not.

NPLOTS: Number of plots in sample (1-5)

TOTN: Total number of taxa in sample

XTOTABCOV: Sum of XABCOV across all taxa in sample

sXRCOV: taxon mean relative cover (XABCOV/XTOTABCOV)*100

FREQ: Relative number of plots in which taxon occurs (NUM/NPLOTS)*100

TOTFREQ: Sum of FREQ across all taxa in sample

SRFREQ: Relative frequency of taxon relative to total frequency (FREQ/TOTFREQ)*100

This data frame is also merged with the input taxa data frames and contains, in addition, GENUS and FAMILY, but also contains (depending on the input taxa lists): CATEGORY, DURATION, GROWTH_HABIT, NWCA_CC, NWCA_NATSTAT, WIS.

byPlotspp: Data frame with data summarized by sampID variables, PLOT, and TAXON at the species level. Each data frame contains:

sampID Variables identified in sampID argument

PLOT: Plot number

state: State of sample location

coeReg: USAC region code

cValReg: C-value region code

variable name provided in taxon_name: Taxon name, consistent with name in vascIn data frame.

COVER: Sum of cover by TAXON within plot

DISTINCT: Distinctness of taxon, value of 1 assigned to each row

byUIDgen: Data frame with data summarized by sampID variables and TAXON at the genus level and contains sampID, state, coeReg, cValReg, TAXON, NUM, XABCOV, and DISTINCT. NUM is the number of plots in which taxon occurs, and XABCOV is the mean absolute COVER across plots. DISTINCT is the value 1 assigned to each row.

byPlotgen: Data frame with data summarized by sampID variables, PLOT, and TAXON at the genus level. Each data frame contains sampID, PLOT, state, coeReg, cValReg, TAXON, COVER, and DISTINCT. DISTINCT assigns the value for each taxon as 1 if the taxon has COVER>0 or 0 if not. COVER is the sum of the COVER variable.

byUIDfam: Data frame with data summarized by sampID variables and TAXON at the family level and contains sampID, state, coeReg, cValReg, TAXON, NUM, XABCOV, and DISTINCT. NUM is the number of plots in which taxon occurs, and XABCOV is the mean absolute COVER across plots. DISTINCT is the value 1 assigned to each row.

byPlotfam: Data frame with data summarized by sampID , PLOT, and TAXON at the family level. Each data frame contains sampID, PLOT, state, coeReg, cValReg, TAXON, COVER, and DISTINCT. DISTINCT assigns the value for each taxon as 1 if the taxon has COVER>0 or 0 if not. COVER is the sum of the COVER variable.

Details

This function calls the createDFs() function, which sums cover by sampID variables, PLOT, TAXON, with sums > 100 truncated to 100 percent.

References

US Environmental Protection Agency. 2016. National Wetland Condition Assessment: 2011 Technical Report. EPA-843-R-15-006. US Environmental Protection Agency, Washington, DC.

Author

Karen Blocksom blocksom.karen@epa.gov

Examples

head(VascPlantEx)
#>   UID STATE USAC_REGION PLOT             USDA_NAME COVER
#> 1   7    FL  CSTL_PLAIN    2 SAGITTARIA LANCIFOLIA    28
#> 2   7    FL  CSTL_PLAIN    3 SAGITTARIA LANCIFOLIA     6
#> 3   7    FL  CSTL_PLAIN    4 SAGITTARIA LANCIFOLIA     5
#> 4   7    FL  CSTL_PLAIN    5 SAGITTARIA LANCIFOLIA    20
#> 5   7    FL  CSTL_PLAIN    5       SPARTINA PATENS     0
#> 6   7    FL  CSTL_PLAIN    1      PANICUM VIRGATUM     0
prepEx <- prepareData(VascPlantEx,
  taxon_name = "USDA_NAME", state = "STATE",
  coeReg = "USAC_REGION", cValReg = "STATE"
)
#> [1] "Missing key variables! Need at least USDA_NAME, FAMILY, GENUS to calculate metrics."

str(prepEx)
#>  NULL