NWCA 2011 Vascular Plant Taxa List

A dataset containing taxonomy and basic characteristics for vascular plant taxa as used in NWCA 2011.

taxaNWCA

Format

A data frame with 3814 observations on the following 12 variables.

SPECIES_NAME_ID

A numeric vector containing the NWCA taxonomic ID number

USDA_NAME

Accepted USDA PLANTS name, or the official NWCA name where not available

DIVISION

Division level of taxonomy

ORDER

Order level of taxonomy

FAMILY

Family level of taxonomy

GENUS

Genus level of taxonomy

SPECIES

Species level of taxonomy

SUBSPECIES

Subspecies level of taxonomy

VARIETY

Variety level of taxonomy

CATEGORY

Plant category of taxon, including MONOCOT, GYMNOSPERM, DICOT, FERN, LICHEN, HORSETAIL, MOSS, LYCOPOD, and LIVERWORT

GROWTH_HABIT

Growth habit of taxon

DURATION

Lifecycle duration of taxon

@note This dataset is the taxa list used for state-specific CC and native status values used in NWCA 2011.

Examples

head(taxaNWCA)
#>   SPECIES_NAME_ID                        USDA_NAME      DIVISION     ORDER
#> 1              58             xELYHORDEUM MACOUNII MAGNOLIOPHYTA CYPERALES
#> 2             135                   ABIES BALSAMEA CONIFEROPHYTA   PINALES
#> 3             142                   ABIES CONCOLOR CONIFEROPHYTA   PINALES
#> 4             146                    ABIES GRANDIS CONIFEROPHYTA   PINALES
#> 5             153                 ABIES LASIOCARPA CONIFEROPHYTA   PINALES
#> 6             155 ABIES LASIOCARPA VAR. LASIOCARPA CONIFEROPHYTA   PINALES
#>     FAMILY      GENUS    SPECIES SUBSPECIES    VARIETY   CATEGORY GROWTH_HABIT
#> 1  POACEAE ELYHORDEUM   MACOUNII                          MONOCOT    GRAMINOID
#> 2 PINACEAE      ABIES   BALSAMEA                       GYMNOSPERM         TREE
#> 3 PINACEAE      ABIES   CONCOLOR                       GYMNOSPERM         TREE
#> 4 PINACEAE      ABIES    GRANDIS                       GYMNOSPERM         TREE
#> 5 PINACEAE      ABIES LASIOCARPA                       GYMNOSPERM         TREE
#> 6 PINACEAE      ABIES LASIOCARPA            LASIOCARPA GYMNOSPERM         TREE
#>    DURATION
#> 1 PERENNIAL
#> 2 PERENNIAL
#> 3 PERENNIAL
#> 4 PERENNIAL
#> 5 PERENNIAL
#> 6 PERENNIAL
str(taxaNWCA)
#> 'data.frame':	3814 obs. of  12 variables:
#>  $ SPECIES_NAME_ID: int  58 135 142 146 153 155 156 290 297 312 ...
#>  $ USDA_NAME      : chr  "xELYHORDEUM MACOUNII" "ABIES BALSAMEA" "ABIES CONCOLOR" "ABIES GRANDIS" ...
#>  $ DIVISION       : chr  "MAGNOLIOPHYTA" "CONIFEROPHYTA" "CONIFEROPHYTA" "CONIFEROPHYTA" ...
#>  $ ORDER          : chr  "CYPERALES" "PINALES" "PINALES" "PINALES" ...
#>  $ FAMILY         : chr  "POACEAE" "PINACEAE" "PINACEAE" "PINACEAE" ...
#>  $ GENUS          : chr  "ELYHORDEUM" "ABIES" "ABIES" "ABIES" ...
#>  $ SPECIES        : chr  "MACOUNII" "BALSAMEA" "CONCOLOR" "GRANDIS" ...
#>  $ SUBSPECIES     : chr  "" "" "" "" ...
#>  $ VARIETY        : chr  "" "" "" "" ...
#>  $ CATEGORY       : chr  "MONOCOT" "GYMNOSPERM" "GYMNOSPERM" "GYMNOSPERM" ...
#>  $ GROWTH_HABIT   : chr  "GRAMINOID" "TREE" "TREE" "TREE" ...
#>  $ DURATION       : chr  "PERENNIAL" "PERENNIAL" "PERENNIAL" "PERENNIAL" ...