NWCA 2016 Vascular Plant Taxa List

A dataset containing taxonomy and basic characteristics for vascular plant taxa as used in NWCA 2016.

taxaNWCA_2016

Format

A data frame with 5044 observations on the following 28 variables.

SPECIES_NAME_ID

A numeric vector containing the NWCA taxonomic ID number

NWCA_NAME

Accepted USDA PLANTS name, or the official NWCA name where not available

ACCEPTED_SYMBOL

Accepted symbol for taxon from USDA PLANTS or NWCA 2016 if not in USDA PLANTS

TAXON_LEVEL

Taxonomic level of name: CL (class), F (Family), G (Genus), GH (Growth habit), H (), NV (Nonvascular), SC (), SF (), SSP (Subspecies), SP (Species), T (), U (), VAR (Variety)

DIVISION

Division level of taxonomy

CLASS

Class level of taxonomy

SUBCLASS

Subclass level of taxonomy

ORDER

Order level of taxonomy

FAMILY

Family level of taxonomy

GENUS

Genus level of taxonomy

CATEGORY

Plant category of taxon, including MONOCOT, GYMNOSPERM, DICOT, FERN, LICHEN, HORSETAIL, MOSS, LYCOPOD, and LIVERWORT

GROWTH_HABIT

Growth habit of taxon as provided by USDA PLANTS

DURATION

Lifecycle duration of taxon as provided by USDA PLANTS

VINE_ALL

Indicator value for all vine types (1/0)

HERB

Indicator value for herbaceous vascular plant (1/0)

SHRUB_COMB

Indicator value for all types of shrubs (1/0)

TREE_COMB

Indicator value for all types of trees (1/0)

@note This dataset is the taxa list used for state-specific CC and native status values used in NWCA 2016.

Examples

head(taxaNWCA_2016)
#>   SPECIES_NAME_ID            NWCA_NAME ACCEPTED_SYMBOL TAXON_LEVEL
#> 1              58 xELYHORDEUM MACOUNII           ELMA8           H
#> 2             129                ABIES           ABIES           G
#> 3             135       ABIES BALSAMEA            ABBA          SP
#> 4             142       ABIES CONCOLOR            ABCO          SP
#> 5             146        ABIES GRANDIS            ABGR          SP
#> 6             153     ABIES LASIOCARPA            ABLA          SP
#>        DIVISION      CLASS     SUBCLASS     ORDER   FAMILY      GENUS
#> 1 MAGNOLIOPHYTA LILIOPSIDA COMMELINIDAE CYPERALES  POACEAE ELYHORDEUM
#> 2 CONIFEROPHYTA  PINOPSIDA                PINALES PINACEAE      ABIES
#> 3 CONIFEROPHYTA  PINOPSIDA                PINALES PINACEAE      ABIES
#> 4 CONIFEROPHYTA  PINOPSIDA                PINALES PINACEAE      ABIES
#> 5 CONIFEROPHYTA  PINOPSIDA                PINALES PINACEAE      ABIES
#> 6 CONIFEROPHYTA  PINOPSIDA                PINALES PINACEAE      ABIES
#>     CATEGORY GROWTH_HABIT  DURATION VINE_ALL HERB SHRUB_COMB TREE_COMB
#> 1    MONOCOT    GRAMINOID PERENNIAL        0    1          0         0
#> 2 GYMNOSPERM         TREE PERENNIAL        0    0          0         1
#> 3 GYMNOSPERM         TREE PERENNIAL        0    0          0         1
#> 4 GYMNOSPERM         TREE PERENNIAL        0    0          0         1
#> 5 GYMNOSPERM         TREE PERENNIAL        0    0          0         1
#> 6 GYMNOSPERM         TREE PERENNIAL        0    0          0         1
str(taxaNWCA_2016)
#> 'data.frame':	5044 obs. of  17 variables:
#>  $ SPECIES_NAME_ID: int  58 129 135 142 146 153 155 156 163 290 ...
#>  $ NWCA_NAME      : chr  "xELYHORDEUM MACOUNII" "ABIES" "ABIES BALSAMEA" "ABIES CONCOLOR" ...
#>  $ ACCEPTED_SYMBOL: chr  "ELMA8" "ABIES" "ABBA" "ABCO" ...
#>  $ TAXON_LEVEL    : chr  "H" "G" "SP" "SP" ...
#>  $ DIVISION       : chr  "MAGNOLIOPHYTA" "CONIFEROPHYTA" "CONIFEROPHYTA" "CONIFEROPHYTA" ...
#>  $ CLASS          : chr  "LILIOPSIDA" "PINOPSIDA" "PINOPSIDA" "PINOPSIDA" ...
#>  $ SUBCLASS       : chr  "COMMELINIDAE" "" "" "" ...
#>  $ ORDER          : chr  "CYPERALES" "PINALES" "PINALES" "PINALES" ...
#>  $ FAMILY         : chr  "POACEAE" "PINACEAE" "PINACEAE" "PINACEAE" ...
#>  $ GENUS          : chr  "ELYHORDEUM" "ABIES" "ABIES" "ABIES" ...
#>  $ CATEGORY       : chr  "MONOCOT" "GYMNOSPERM" "GYMNOSPERM" "GYMNOSPERM" ...
#>  $ GROWTH_HABIT   : chr  "GRAMINOID" "TREE" "TREE" "TREE" ...
#>  $ DURATION       : chr  "PERENNIAL" "PERENNIAL" "PERENNIAL" "PERENNIAL" ...
#>  $ VINE_ALL       : chr  "0" "0" "0" "0" ...
#>  $ HERB           : chr  "1" "0" "0" "0" ...
#>  $ SHRUB_COMB     : chr  "0" "0" "0" "0" ...
#>  $ TREE_COMB      : chr  "0" "1" "1" "1" ...