NWCA 2016 Wetland Indicator Status

A dataset containing NWCA 2016 USAC-specific Wetland Indicator Status values for vascular plant taxa

wisNWCA_2016

Format

A data frame with 9584 observations on the following 5 variables.

SPECIES_NAME_ID

A numeric vector containing the NWCA taxonomic ID number

NWCA_NAME

Accepted USDA PLANTS name, or the official NWCA name where not available

GEOG_ID

Geographic unit to which value applies, which is US Army Corps region in this dataset

WIS

USAC-specific Wetland Indicator Status as used for NWCA 2016

ECOIND1

Wetland indicator status converted to numeric value, with lower numbers more characteristic of wetlands

ECOIND2

Wetland indicator status converted to numeric value, with higher numbers more characteristic of wetlands

Note

This dataset is the taxa list used for USAC-specific Wetland Indicator Status values used in NWCA 2016.

Examples

head(wisNWCA_2016)
#>   SPECIES_NAME_ID            NWCA_NAME   GEOG_ID WIS ECOIND1 ECOIND2
#> 1              58 xELYHORDEUM MACOUNII GT_PLAINS FAC       3       3
#> 2             129                ABIES    E_MTNS UND      NA      NA
#> 3             135       ABIES BALSAMEA    E_MTNS FAC       3       3
#> 4             135       ABIES BALSAMEA        NE FAC       3       3
#> 5             142       ABIES CONCOLOR    ARID_W UPL       5       1
#> 6             142       ABIES CONCOLOR    W_MTNS UPL       5       1
str(wisNWCA_2016)
#> 'data.frame':	9584 obs. of  6 variables:
#>  $ SPECIES_NAME_ID: int  58 129 135 135 142 142 146 153 153 155 ...
#>  $ NWCA_NAME      : chr  "xELYHORDEUM MACOUNII" "ABIES" "ABIES BALSAMEA" "ABIES BALSAMEA" ...
#>  $ GEOG_ID        : chr  "GT_PLAINS" "E_MTNS" "E_MTNS" "NE" ...
#>  $ WIS            : chr  "FAC" "UND" "FAC" "FAC" ...
#>  $ ECOIND1        : int  3 NA 3 3 5 5 4 4 4 4 ...
#>  $ ECOIND2        : int  3 NA 3 3 1 1 2 2 2 2 ...